% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R, R/RaspSolved.R, R/generics.R
\name{amount.target.RaspData}
\alias{amount.target}
\alias{amount.target.RaspData}
\alias{amount.target.RaspUnsolOrSol}
\alias{amount.target<-}
\alias{amount.target<-.RaspData}
\alias{amount.target<-.RaspUnsolOrSol}
\title{Amount targets}
\usage{
\method{amount.target}{RaspData}(x, species = NULL)

\method{amount.target}{RaspData}(x, species = NULL) <- value

\method{amount.target}{RaspUnsolOrSol}(x, species = NULL)

\method{amount.target}{RaspUnsolOrSol}(x, species = NULL) <- value

amount.target(x, species)

amount.target(x, species) <- value
}
\arguments{
\item{x}{\code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.}

\item{species}{\code{NULL} for all species or \code{integer} indicating species.}

\item{value}{\code{numeric} new target.}
}
\value{
code{numeric} vector.
}
\description{
This function sets or returns the target amounts for each species.
}
\examples{
# load data
data(sim_rs)
# extract amount targets for all species
amount.target(sim_rs)
# set amount targets for all species
amount.target(sim_rs) <- 0.1
# extract amount targets for first species
amount.target(sim_rs, 1)
# set amount targets for for first species
amount.target(sim_rs, 1) <- 0.5
}
\seealso{
\code{RaspData}, \code{RaspResults}, \code{RaspSolved}.
}

