% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R, R/RaspSolved.R, R/generics.R
\name{amount.targets.met.RaspResults}
\alias{amount.targets.met}
\alias{amount.targets.met.RaspResults}
\alias{amount.targets.met.RaspSolved}
\title{Extract information on whether solutions have met the amount targets}
\usage{
\method{amount.targets}{met.RaspResults}(x, y = NULL)

\method{amount.targets}{met.RaspSolved}(x, y = NULL)

amount.targets.met(x, ...)
}
\arguments{
\item{x}{\code{RaspResults} or \code{RaspSolved} object.}

\item{y}{\code{NULL} to return all values, \code{integer} 0 to return values for best solution, \code{integer} value greater than 0 for \code{y}'th solution value.}

\item{...}{not used.}
}
\value{
\code{matrix} or \code{numeric} vector depending on arguments.
}
\description{
This function reports whether a solution has met the amount targets for each species in a solution.
}
\seealso{
\code{\link{RaspResults-class}}, \code{\link{RaspSolved}}, \code{\link{rasp}}
}

