% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{amount.targets}
\alias{amount.targets}
\title{Extract amount targets}
\usage{
amount.targets(x, ...)
}
\arguments{
\item{x}{any \code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.}

\item{...}{not used.}
}
\description{
This function returns the amount targets (%) for species in a RASP object.
}
\note{
This generic method does not work on \code{RaspResults} objects because they do not store this information.
}
\seealso{
\code{\link{RaspOpts-class}}, \code{\link{RaspData-class}}, \code{\link{RaspUnsolved-class}}, \code{\link{RaspSolved-class}}
}

