% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GurobiOpts.R, R/generics.R
\name{as.list.GurobiOpts}
\alias{as.list}
\alias{as.list.GurobiOpts}
\title{Convert object to list.}
\usage{
\method{as.list}{GurobiOpts}(x, ...)
}
\arguments{
\item{x}{\code{GurobiOpts} object.}

\item{...}{not used.}
}
\value{
\code{list}
}
\description{
Convert \code{GurobiOpts} object to list.
}
\note{
This function will not include the \code{NumberSolutions} slot or the \code{TimeLimit} slot if it is not finite.
}
\examples{
# make GuboriOpts object
x <- GurobiOpts()
# convert to list
as.list(x)
}
\seealso{
\code{GurobiOpts}.
}

