% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AttributeSpace.R
\docType{class}
\name{AttributeSpace-class}
\alias{AttributeSpace-class}
\title{AttributeSpace: An S4 class to represent an attribute space.}
\description{
This class is used to store planning unit points and demand points in an attribute space.
}
\section{Slots}{

\describe{
\item{\code{dp}}{\code{SimplePoints} coordinates for planning unit in the space.}

\item{\code{pu}}{\code{list} of \code{DemandPoints} for each species.}
}}
\seealso{
\code{\link{SimplePoints-class}}, \code{\link{DemandPoints-class}}.
}

