% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AttributeSpace.R
\name{AttributeSpace}
\alias{AttributeSpace}
\title{Create new AttributeSpace object}
\usage{
AttributeSpace(pu, dp)
}
\arguments{
\item{pu}{\code{SimplePoints} coordinates for planning unit in the space.}

\item{dp}{\code{list} of \code{DemandPoints} for each species.}
}
\description{
This function creates a new \code{AttributeSpace} object.
}
\examples{
space <- AttributeSpace(
	pu <- SimplePoints(
		matrix(rnorm(100), ncol=2)
	),
	dp <- list(
	DemandPoints(
			SimplePoints(matrix(rnorm(100), ncol=2)),
			runif(50)
		)
	)
)
}
\seealso{
\code{\link{SimplePoints-class}}, \code{\link{DemandPoints-class}}.
}

