% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttributeSpaces.R
\docType{class}
\name{AttributeSpaces-class}
\alias{AttributeSpaces-class}
\title{AttributeSpaces: An S4 class to represent a collection of attribute spaces for different species.}
\description{
This class is used to store a collection of attribute spaces for different species.
}
\section{Slots}{

\describe{
\item{\code{spaces}}{\code{list} of \code{\link{AttributeSpace}} objects for different species.}

\item{\code{name}}{\code{character} name to identify the attribute space.}
}}

\seealso{
\code{\link{AttributeSpace-class}}.
}
