% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{blank.raster}
\alias{blank.raster}
\title{Blank raster}
\usage{
blank.raster(x, res)
}
\arguments{
\item{x}{\code{Spatial*} object.}

\item{res}{\code{numeric vector} specifying resolution of the output raster in the x and y dimensions. If \code{vector} is of length one, then the pixels are assumed to be square.}
}
\description{
This functions creates a blank raster based on the spatial extent of a Spatial object.
}
\examples{
# load data
data(sim_pus)
# make RasterLayer from SpatialPolygons
blank.raster(sim_pus, 1)
}

