% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RapData.R, R/RapResults.R, R/rapr-internal.R
\docType{methods}
\name{cache,RapData,character,ANY-method}
\alias{cache}
\alias{cache,RapData,character,ANY-method}
\alias{cache,RapData,character,missing-method}
\alias{cache,RapResults,character,ANY-method}
\alias{cache,RapResults,character,missing-method}
\title{Get and set cache Methods}
\usage{
\S4method{cache}{RapData,character,ANY}(x, name, y)

\S4method{cache}{RapData,character,missing}(x, name, y)

\S4method{cache}{RapResults,character,ANY}(x, name, y)

\S4method{cache}{RapResults,character,missing}(x, name, y)

cache(x, name, y)
}
\arguments{
\item{x}{\code{RapData} or \code{RapResults} object}

\item{name}{\code{character} hash.}

\item{y}{if \code{ANY} this object gets cached with name, else if \code{missing} the object hashed at name gets returned.}
}
\value{
\code{ANY} or \code{NULL} depends on \code{y} argument.
}
\description{
Getter and setter methods for caches in RapData and RapResults object.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = RapData,name = character,y = ANY}: 

\item \code{x = RapData,name = character,y = missing}: 

\item \code{x = RapResults,name = character,y = ANY}: 

\item \code{x = RapResults,name = character,y = missing}: 
}}
\note{
caches are implemented using environments, the hash is used as the name of the object in the environment.
}
\keyword{internal}

