% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{casestudy_data}
\alias{casestudy_data}
\alias{cs_pus}
\title{Case-study dataset for a conservation planning exercise}
\description{
This dataset contains data to generate example prioritisations for the pale-headed Rosella (\textit{Platycercus adscitus}). Specific objects in the dataset include:
\itemize{
	\item cs_spp: \code{RasterLayer} probability distribution map for the \textit{P. adscitus} dervived from records in the BirdLife Altas Database (\url{http://birdata.com.au/}).
	\item cs_pus: \code{SpatialPolygonsDataFrame} planning units. The units were generated as $10km^2$ squares across the species' range, and then clipped to the coast-line. They were then overliad with Australia's protected area network (obtained from the World Database on Protected Areas (WDPA) at \url{http://www.protectedplanet.net/}). This attribute table has 3 fields. The 'area' field contains the units' area, the 'cost' field is set to 1 for all units, and the 'status' filed indicates if 50\% or moer of the units' extent is covered by protected areas.
	\item cs_bio12: \code{RasterLayer} containing annual precipitation data across the species range (obtained from \url{http://www.worldclim.org/} as BIO12, and resampled ot $10km^2$ resolution).
}
}
\examples{
# load data
data(cs_pus)
data(cs_spp)
data(cs_bio12)
# plot data
spplot(cs_pus, field=status)
plot(cs_spp)
plot(cs_bio12)
}
\keyword{datasets}

