% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray.R
\name{collapse_probe_ids}
\alias{collapse_probe_ids}
\title{Collapse a expression of probes to gene symbols}
\usage{
collapse_probe_ids(accesion_number)
}
\arguments{
\item{accesion_number}{}
}
\value{

}
\description{
Collapse a expression of probes to gene symbols
}
