% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corner.R
\name{corner}
\alias{corner}
\title{Returns the first or last rows and columns of a subsettable object (matrix, data frame, tibble, data table). Alternative to head or tail.}
\usage{
corner(x, m = 5, n = 5)
}
\arguments{
\item{x}{An object}

\item{m}{a single positive integer. The number of rows to return.}

\item{n}{a single positive integer. The number of columns to return.}
}
\description{
Returns the first or last rows and columns of a subsettable object (matrix, data frame, tibble, data table). Alternative to head or tail.
}
