% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demand.points.R
\name{demand.points}
\alias{demand.points}
\title{Generate demand points for RASP}
\usage{
demand.points(species.points, space.rasters = NULL, id = 1L,
  demand.points = 1000L, kernel.method = c("sm.density", "hypervolume")[1],
  ...)
}
\arguments{
\item{species.points}{\code{SpatialPointsDataFrame} or \code{SpatialPoints} with species presence records.}

\item{space.rasters}{\code{NULL}, \code{RasterLayer}, \code{RasterStack}, \code{RasterBrick} with projections of the attribute space over geographic space. If NULL then geographic space is used.}

\item{id}{\code{integer} species id.}

\item{demand.points}{\code{integer} number of demand points to use for each attribute space for each species.}

\item{kernel.method}{\code{character} name of kernel method to use to generate demand points.}

\item{...}{arguments passed to kernel density estimating functions}
}
\value{
\code{DemandPoints} object.
}
\description{
This function generates demand points for RASP using kernels.
}
\details{
The values of the species records in the rasters is extracted. A kernel is fit to the points in the attribute space.
Volumes are then fit to the points in the attribute space(s). Points are randomly generated inside the volume. Demand points are generated
as random points inside the volume. A kernel is fit to the species records and used to predict the density at each of the demand points.
By using 'sm.density' as an argument in \code{kernel.method}, the volume is fit as a minimum convex polygon using \code{\link[rgeos]{gConvex}} and
\code{\link[sm]{sm.density}} is used to fit the kernel. Note this can only be used when the data is low-dimensional (d<3). By using
}
\seealso{
\code{\link[hypervolume]{hypervolume}}, \code{\link[sm]{sm.density}}, \code{\link[rgeos]{gConvex}}.
}

