% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DemandPoints.R
\docType{class}
\name{DemandPoints-class}
\alias{DemandPoints-class}
\title{DemandPoints: An S4 class to represent demand points}
\description{
This class is used to store demand point information.
}
\section{Slots}{

\describe{
\item{\code{coords}}{\code{matrix} coordinates for each demand point in the attribute space.}

\item{\code{weights}}{\code{numeric} weights for each demand poont.}

\item{\code{species}}{\code{integer} idenifiers for which species each demand point belongs to.}
}}
\seealso{
\code{\link{DemandPoints}}
}

