% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DemandPoints.R
\name{DemandPoints}
\alias{DemandPoints}
\title{Create new DemandPoints object}
\usage{
DemandPoints(coords, weights, species)
}
\description{
This funciton creates a new DemandPoints object
}
\section{Slots}{

\describe{
\item{\code{coords}}{\code{matrix} coordinates for each demand point in the attribute space.}

\item{\code{weights}}{\code{numeric} weights for each demand poont.}

\item{\code{species}}{\code{integer} idenifiers for which species each demand point belongs to.}
}}
\seealso{
\code{\link{DemandPoints-class}}
}

