% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{downloadGEO}
\alias{downloadGEO}
\title{Download files associated with a GEO accession number}
\usage{
downloadGEO(accessionNumber, dataDirectory = NULL)
}
\arguments{
\item{accessionNumber}{GEO accession number of study}

\item{dataDirectory}{path to directory to place files}
}
\value{

}
\description{
Download files associated with a GEO accession number
}
