% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R, R/RaspSolved.R, R/RaspUnsolved.R, R/generics.R
\name{dp.subset.RaspData}
\alias{dp.subset}
\alias{dp.subset.RaspData}
\alias{dp.subset.RaspSolved}
\alias{dp.subset.RaspUnsolved}
\title{Subset demand points}
\usage{
\method{dp.subset}{RaspData}(x, space, species, points)

\method{dp.subset}{RaspSolved}(x, space, species, points)

\method{dp.subset}{RaspUnsolved}(x, space, species, points)

dp.subset(x, space, species, points)
}
\arguments{
\item{x}{\code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.}

\item{space}{\code{integer} vector to specify the index of the space to subset demand points from.}

\item{species}{\code{integer} vector to specify the index of the species to subset demand points from.}

\item{points}{\code{integer} vector to specify the index of demand points to subset.}
}
\value{
\code{RaspData} or \code{RaspUnsolved} object depending on input object.
}
\description{
Subset demand points from a \code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.
}
\examples{
# load data
data(sim_ru)
# generate new object with first 10 planning units
sim_ru2 <- dp.subset(sim_ru, 1, 1, 1:10)
}
\seealso{
\code{RaspData}, \code{RaspUnsolved}, \code{RaspSolved}.
}

