% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R
\name{format.RaspData}
\alias{format.RaspData}
\title{Format data for RASP}
\usage{
format.RaspData(polygons, species.points, species.rasters, space.rasters, area.targets = 0.2,
space.targets = 0.2,  demand.points=1000, kernel.method='guassian', include.geographic.space=TRUE,
pu = NULL, species = NULL, species.coords = NULL, pu.species.probabilities = NULL,
boundary = NULL, ..., verbose = FALSE)
}
\arguments{
\item{polygons}{\code{SpatialPolygons} with planning unit data.}

\item{species.rasters}{\code{RasterLayer}, \code{RasterStack}, \code{RasterBrick} with species probability distribution data.}

\item{space.rasters}{\code{list} of/or \code{RasterLayer}, \code{RasterStack}, \code{RasterBrick} representing projects of attribute space over geographic space. Use a \code{list} to denote seperate attribute spaces.}

\item{area.targets}{\code{numeric} vector for area targets (%) for each species. Defaults to 0.2 for each attribute space for each species.}

\item{space.targets}{\code{numeric} vector for attribute space targets (%) for each species. Defaults to 0.2 for each attribute space for each species. Note all attribute spaces have the same targets.}

\item{kernel.method}{\code{character} name of kernel method to use to generate demand points. Use either \code{'sm.density'} or \code{'hypervolume'}.}

\item{species.points}{\code{list} of/or \code{SpatialPointsDataFrame} or \code{SpatialPoints} with species presence records. Use a \code{list} of objects to represent different species. Must have the same number of elements as \code{species.rasters}.}

\item{...}{additional arguments to \code{calcBoundaryData} and \code{calcPuVsSpeciesData}.}

\item{verbose}{\code{logical} print statements during processing?}

\item{n.demand.points}{\code{integer} number of demand points to use for each attribute space for each species.}

\item{include.geographic.space}{\code{logical} should the geographic space be considered an attribute space?}
}
\description{
This function prepares spatially explicit planning unit and species data for RASP processing.
By default, raw planning unit and species distribution data can be used to generate all RASP inputs from scratch.
}
\examples{
data(pus, species.points, species.rasters, space.rasters, demand.points)
x<-RaspData(pus, species.rasters=species.rasters, space.rasters=space.rasters, demand.points=demand.points)
}
\seealso{
\code{\link{RaspData-class}}, \code{\link{RaspData}}.
}

