% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{geneSetEnrichment}
\alias{geneSetEnrichment}
\title{Perform gene set enrichment analysis}
\usage{
geneSetEnrichment(geneSets, expressionVals, genes, minSetSize = 0)
}
\arguments{
\item{geneSets}{a list of gene name vectors}

\item{expressionVals}{a vector of gene scores such as differential expression values}

\item{genes}{a vector of gene names in the same order as expressionVals}

\item{minSetSize}{minimum number of genes that must be in gene set after filtering}
}
\value{

}
\description{
Perform gene set enrichment analysis
}
