% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{getFeatureData}
\alias{getFeatureData}
\title{Download feature data associated with a GEO accession number}
\usage{
getFeatureData(accessionNumber)
}
\arguments{
\item{accessionNumber}{GEO accession number of study}
}
\value{

}
\description{
Download feature data associated with a GEO accession number
}
