% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_enrichment.R
\name{getMaxScores}
\alias{getMaxScores}
\title{Return tibble with max enrichment score for each gene set}
\usage{
getMaxScores(enrichmentScores)
}
\arguments{
\item{enrichmentScores}{output of geneSetEnrichment}
}
\value{

}
\description{
Return tibble with max enrichment score for each gene set
}
