% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{getMetaData}
\alias{getMetaData}
\title{Download meta data associated with a GEO accession number}
\usage{
getMetaData(accessionNumber)
}
\arguments{
\item{accessionNumber}{GEO accession number of study}
}
\value{

}
\description{
Download meta data associated with a GEO accession number
}
