% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{getProbeData}
\alias{getProbeData}
\title{Download probe data associated with a GEO accession number}
\usage{
getProbeData(accessionNumber)
}
\arguments{
\item{accessionNumber}{GEO accession number of study}
}
\value{

}
\description{
Download probe data associated with a GEO accession number
}
