% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GurobiOpts.R
\name{GurobiOpts}
\alias{GurobiOpts}
\title{Create GurobiOpts object}
\usage{
GurobiOpts(NTHREADS = 1L, GAP = 0.05, PRESOLVE = 2L,
  TIMELIMIT = NA_integer_, VERBOSITY = 1L)
}
\arguments{
\item{NTHREADS}{\code{integer} number of cores to use for processing. Defaults to 1L.}

\item{GAP}{\code{numeric} MIP gap (lp_solve parameter). Defaults to 0.05.}

\item{PRESOLVE}{\code{integer} code for level of computation in presolve (lp_solve parameter). Defaults to 2.}

\item{TIMELIMIT}{\code{integer} number of seconds to allow for solving. Defaults to NA_integer_, and so a time limit is not imposed.}

\item{VERBOSITY}{\code{integer} code for amount of output to display. Defaults to 1L.}
}
\value{
\code{GurobiOpts} object
}
\description{
This function creates a new GurobiOpts object.
}
\examples{
x<-GurobiOpts(NTHREADS=2)
}
\seealso{
\code{\link{GurobiOpts-class}}.
}

