% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_theme.R
\name{heatmap_color}
\alias{heatmap_color}
\title{Title}
\usage{
heatmap_color(
  colors = c("#0A95A7", "white", "#E5541B"),
  range = c(-1, 0, 1),
  space = "LAB"
)
}
\arguments{
\item{space}{}
}
\value{

}
\description{
Title
}
