% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray.R
\name{import_CEL}
\alias{import_CEL}
\title{import a CEL file}
\usage{
import_CEL(directory, probe_file = "../data/GPL570-55999.txt", skip_lines = 16)
}
\arguments{
\item{directory}{path to directory with CEL files}

\item{probe_file}{provide path to the probe file}

\item{skip_lines}{number of lines to skip in probe file}
}
\value{

}
\description{
import a CEL file
}
