% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R, R/RaspResults.R, R/raspr-internal.R
\docType{methods}
\name{is.cached,RaspData,character-method}
\alias{is.cached}
\alias{is.cached,RaspData,character-method}
\alias{is.cached,RaspResults,character-method}
\title{Test if hash is cached in a Rasp object}
\usage{
\S4method{is.cached}{RaspData,character}(x, name)

\S4method{is.cached}{RaspResults,character}(x, name)

is.cached(x, name)
}
\arguments{
\item{x}{\code{RaspData} or \code{RaspResults} object}

\item{name}{\code{character} hash.}
}
\value{
\code{logical} Is it cached?
}
\description{
Tests if hash is cached in Rasp object.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = RaspData,name = character}: 

\item \code{x = RaspResults,name = character}: 
}}
\note{
caches are implemented using environments, the hash is used as the name of the object in the environment.
}
\keyword{internal}

