% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{is.GurobiInstalled}
\alias{is.GurobiInstalled}
\title{Test if Gurobi is installed on computer}
\usage{
is.GurobiInstalled()
}
\value{
"logical" Is it installed and ready to use?
}
\description{
This function determines if Gurobi is installed on the computer, check that its licensing is set up, and will store the path in \code{\link[base]{options}}.
}
\examples{
is.GurobiInstalled()
options()$GurobiInstalled
}
\seealso{
\code{\link[base]{options}}.
}

