% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_enrichment.R
\name{kernelEnrichment}
\alias{kernelEnrichment}
\title{Perform gene set enrichment analysis using a kernel-weighted GSEA score on a vector of values for a single sample}
\usage{
kernelEnrichment(
  geneSets,
  expressionVals,
  genes,
  minSetSize = 0,
  randomization_n = 1000,
  parallel = FALSE
)
}
\arguments{
\item{geneSets}{a list of gene name vectors}

\item{expressionVals}{a vector of gene scores such as differential expression values or Z-scores}

\item{genes}{a vector of gene names in the same order as expressionVals}

\item{minSetSize}{minimum number of genes that must be in gene set after filtering}

\item{randomization_n}{number of rounds of randomization for empirical p-value calculation}

\item{parallel}{run analysis using multicore parallel operations *may break*}
}
\value{
tibble with rows for each geneSet
}
\description{
Perform gene set enrichment analysis using a kernel-weighted GSEA score on a vector of values for a single sample
}
