% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_enrichment.R
\name{kernelEnrichmentMulti}
\alias{kernelEnrichmentMulti}
\title{Perform gene set enrichment analysis using a kernel-weighted GSEA score on a matrix of genes x samples}
\usage{
kernelEnrichmentMulti(
  geneSets,
  expressionVals,
  genes,
  minSetSize = 0,
  randomization_n = 1000,
  parallel = FALSE
)
}
\arguments{
\item{geneSets}{a list of gene name vectors}

\item{expressionVals}{a matrix/table of gene scores such as differential expression values or Z-scores, genes x samples}

\item{genes}{a vector of gene names in the same order as the rows of expressionVals}

\item{minSetSize}{minimum number of genes that must be in gene set after filtering}

\item{randomization_n}{number of rounds of randomization for empirical p-value calculation}

\item{parallel}{run analysis using multicore parallel operations *may break*}
}
\value{
tibble with rows for each geneSet x each sample
}
\description{
Perform gene set enrichment analysis using a kernel-weighted GSEA score on a matrix of genes x samples
}
