% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_enrichment.R
\name{kernelEnrichmentPlotter}
\alias{kernelEnrichmentPlotter}
\title{Plot the scores returned by kernelEnrichment/kernelEnrichmentMulti}
\usage{
kernelEnrichmentPlotter(
  data,
  group,
  plot.title = "Gene Set Enrichment Plot",
  label.lines = TRUE
)
}
\arguments{
\item{data}{a results table returned by kernelEnrichment/kernelEnrichmentMulti}

\item{group}{a column in `data` used to color the plotted lines}

\item{plot.title}{text to use for the plot title}

\item{label.lines}{logical, whether to add labels to each line}
}
\value{

}
\description{
Plot the scores returned by kernelEnrichment/kernelEnrichmentMulti
}
