% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R, R/RaspSolved.R, R/generics.R
\name{log.file.RaspResults}
\alias{log.file}
\alias{log.file.RaspResults}
\alias{log.file.RaspSolved}
\title{Log file}
\usage{
\method{log}{file.RaspResults}(x)

\method{log}{file.RaspSolved}(x)

log.file(x, ...)
}
\arguments{
\item{x}{\code{RaspResults} or \code{RaspSolved} object.}

\item{...}{not used.}
}
\description{
This function returns the Gurobi log file (*.log) associated with solving RASP.
}
\seealso{
\code{\link{RaspResults-class}}, \code{\link{RaspSolved}}, \code{\link{rasp}}
}

