% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R, R/RaspSolved.R, R/generics.R
\name{logging.file.RaspResults}
\alias{logging.file}
\alias{logging.file.RaspResults}
\alias{logging.file.RaspSolved}
\title{Log file}
\usage{
\method{logging.file}{RaspResults}(x)

\method{logging.file}{RaspSolved}(x)

logging.file(x)
}
\arguments{
\item{x}{\code{RaspResults} or \code{RaspSolved} object.}
}
\description{
This function returns the Gurobi log file (*.log) associated with solving RASP.
}
\seealso{
\code{\link{RaspResults-class}}, \code{\link{RaspSolved-class}}, \code{\link{rasp}}.
}

