% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DemandPoints.R
\name{make.DemandPoints}
\alias{make.DemandPoints}
\title{Generate demand points for RASP}
\usage{
make.DemandPoints(species.points, space.rasters = NULL, n = 1000L,
  quantile = 0.2, kernel.method = c("sm.density", "hypervolume")[1], ...)
}
\arguments{
\item{species.points}{\code{SpatialPointsDataFrame} or \code{SpatialPoints} with species presence records.}

\item{space.rasters}{\code{NULL}, \code{RasterLayer}, \code{RasterStack}, \code{RasterBrick} with projections of the attribute space over geographic space. If NULL (default) then geographic space is used.}

\item{n}{\code{integer} number of demand points to use for each attribute space for each species. Defaults to 1000L.}

\item{quantile}{\code{numeric} quantile to generate demand points within. If 0 then demand points are generated across the full range of values the \code{species.points} intersect. Defaults to 0.2.}

\item{kernel.method}{\code{character} name of kernel method to use to generate demand points. Defaults to 'sm.density'.}

\item{...}{arguments passed to kernel density estimating functions}
}
\value{
\code{DemandPoints} object.
}
\description{
This function generates demand points for RASP using kernels.
}
\details{
The values of the species records in the rasters is extracted. A kernel is fit to the points in the attribute space.
Volumes are then fit to the points in the attribute space(s). Points are randomly generated inside the volume. Demand points are generated
as random points inside the volume. A kernel is fit to the species records and used to predict the density at each of the demand points.
By using 'sm.density' as an argument in \code{kernel.method}, the volume is fit as a minimum convex polygon using \code{\link[rgeos]{gConvex}} and
\code{\link[sm]{sm.density}} is used to fit the kernel. Note this can only be used when the data is low-dimensional (d<3). By using
}
\seealso{
\code{\link[hypervolume]{hypervolume}}, \code{\link[sm]{sm.density}}, \code{\link[rgeos]{gConvex}}.
}

