% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManualOpts.R
\name{ManualOpts-class}
\alias{ManualOpts-class}
\title{ManualOpts: An S4 class to represent parameters for manually specified
solutions}
\description{
This class is used to store parameters.
}
\section{Slots}{

\describe{
\item{\code{NumberSolutions}}{\code{integer} number of solutions.}
}}

\seealso{
\code{\link{ManualOpts}}.
}
