% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DemandPoints.R
\name{merge.DemandPoints}
\alias{merge.DemandPoints}
\title{Merge a list of DemandPoints objects togeather}
\usage{
merge.DemandPoints(x, ...)
}
\arguments{
\item{x}{\code{list} of \code{DemandPoints} objects.}
}
\description{
This is a conviencne function to merge DemandPoints objects togeather.
}
\seealso{
\code{DemandPoints}
}

