% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R
\name{merge.RaspResults}
\alias{merge.RaspResults}
\title{Merge RASP results}
\usage{
merge.RaspResults(x)
}
\arguments{
\item{x}{\code{list} of \code{RaspResults} objects.}
}
\value{
\code{RaspResults} object
}
\description{
This function merges a list of \code{RaspResults} objects into a single \code{RaspResults} object.
It is used for collating results from multiple runs.
}
\seealso{
\code{\link{RaspResults-class}}, \code{\link{RaspResults}}.
}

