% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R, R/RaspSolved.R, R/generics.R
\name{model.file.RaspResults}
\alias{model.file}
\alias{model.file.RaspResults}
\alias{model.file.RaspSolved}
\title{Modelfile}
\usage{
\method{model.file}{RaspResults}(x)

\method{model.file}{RaspSolved}(x)

model.file(x, ...)
}
\arguments{
\item{x}{\code{RaspResults} or \code{RaspSolved} object.}

\item{...}{not used.}
}
\description{
This function returns the Gurobi model file associated with solving RASP.
}
\seealso{
\code{\link{RaspResults-class}}, \code{\link{RaspSolved}}, \code{\link{rasp}}
}

