% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/RapData.R, R/RapSolved.R
\name{names}
\alias{names}
\alias{names<-.RapData}
\alias{names.RapData}
\alias{names<-.RapUnsolOrSol}
\alias{names.RapUnsolOrSol}
\title{Names}
\usage{
\method{names}{RapData}(x) <- value

\method{names}{RapData}(x)

\method{names}{RapUnsolOrSol}(x) <- value

\method{names}{RapUnsolOrSol}(x)
}
\arguments{
\item{x}{\code{\link{RapData}}, \code{\link{RapUnsolved}}, or \code{\link{RapSolved}} object.}

\item{value}{new species names.}
}
\description{
This function sets or returns the species names in an object.
}
\examples{
# load data
data(sim_rs)
# show names
names(sim_rs)
# change names
names(sim_rs) <- c('spp1', 'spp2', 'spp3')
# show new names
names(sim_rs)
}
\seealso{
\code{\link{RapData}}, \code{\link{RapUnsolved}}, \code{\link{RapSolved}}.
}
