% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R, R/RaspSolved.R, R/generics.R
\name{occ.held.RaspResults}
\alias{occ.held}
\alias{occ.held.RaspResults}
\alias{occ.held.RaspSolved}
\title{Extract occurrence held for a solution}
\usage{
\method{occ.held}{RaspResults}(x, y = NULL)

\method{occ.held}{RaspSolved}(x, y = NULL)

occ.held(x, ...)
}
\arguments{
\item{x}{\code{RaspResults} or \code{RaspSolved} object.}

\item{y}{\code{NULL} to return all values, \code{integer} 0 to return values for best solution, \code{integer} value greater than 0 for \code{y}'th solution value.}

\item{...}{not used.}
}
\value{
\code{matrix} or \code{numeric} vector depending on arguments.
}
\description{
This function returns the number of occurrences held for each species in a solution.
}
\seealso{
\code{\link{RaspResults-class}}, \code{\link{RaspSolved}}, \code{\link{rasp}}
}

