% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlanningUnitPoints.R
\docType{class}
\name{PlanningUnitPoints-class}
\alias{PlanningUnitPoints-class}
\title{PlanningUnitPoints: An S4 class to represent planning units in an attribute space}
\description{
This class is used to planning units in an attribute space.
}
\section{Slots}{

\describe{
\item{\code{coords}}{\code{\link[base]{matrix}} coordinates for each point.}

\item{\code{ids}}{\code{integer} planning unit ids.}
}}

\seealso{
\code{\link{AttributeSpace}}.
}
