% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlanningUnitPoints.R
\name{PlanningUnitPoints}
\alias{PlanningUnitPoints}
\title{Create new PlanningUnitPoints object}
\usage{
PlanningUnitPoints(coords, ids)
}
\arguments{
\item{coords}{\code{\link[base:matrix]{base::matrix()}} coordinates for each point.}

\item{ids}{\code{integer} planning unit ids.}
}
\description{
This function creates a new \code{PlanningUnitPoints} object.
}
\examples{
# create PlanningUnitPoints object
x <- PlanningUnitPoints(matrix(rnorm(150), ncol = 1), seq_len(150))

# print object
print(x)

}
\seealso{
\linkS4class{AttributeSpace}.
}
