% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RapSolved.R, R/generics.R
\docType{methods}
\name{plot,RapSolved,numeric-method}
\alias{plot}
\alias{plot,RapSolved,RapSolved-method}
\alias{plot,RapSolved,missing-method}
\alias{plot,RapSolved,numeric-method}
\title{Plot RASP object}
\usage{
\S4method{plot}{RapSolved,numeric}(x, y, basemap = "none",
  color.palette = "Greens", locked.in.color = "#000000FF",
  locked.out.color = "#D7D7D7FF", alpha = ifelse(basemap == "none", 1, 0.7),
  grayscale = FALSE, force.reset = FALSE)

\S4method{plot}{RapSolved,missing}(x, y, basemap = "none",
  color.palette = "PuBu", locked.in.color = "#000000FF",
  locked.out.color = "#D7D7D7FF", alpha = ifelse(basemap == "none", 1, 0.7),
  grayscale = FALSE, force.reset = FALSE)

\S4method{plot}{RapSolved,RapSolved}(x, y, i = NULL, j = i,
  basemap = "none", color.palette = ifelse(is.null(i), "RdYlBu", "Accent"),
  x.locked.in.color = "#000000FF", x.locked.out.color = "#D7D7D7FF",
  y.locked.in.color = "#FFFFFFFF", y.locked.out.color = "#D7D7D7FF",
  alpha = ifelse(basemap == "none", 1, 0.7), grayscale = FALSE,
  force.reset = FALSE)
}
\arguments{
\item{x}{\code{RapSolved} object.}

\item{y}{\code{NULL} to plot selection frequencies. \code{numeric} to plot the i'th solution, or 0 to plot the best solution. \code{RapSolved} object to plot differences in solutions between objects. Defaults to \code{ULL}.}

\item{basemap}{\code{character} object indicating the type of basemap to use (see \code{link{basemap}}). Use either 'none', 'roadmap', 'mobile', 'satellite', 'terrain', 'hybrid', 'mapmaker-roadmap', 'mapmaker-hybrid'. Defaults to 'none'.}

\item{color.palette}{\code{character} name of colour palette to use for planning units (see \code{\link[RColorBrewer]{brewer.pal}}). Defaults to 'PuBu' when \code{y} is \code{NULL}, 'Greens' when \code{y} is \code{numeric}, and 'RdYlBu' or 'Accent' when \code{y} is \code{RapSolved}.}

\item{locked.in.color}{\code{character} color to denote locked in planning units. Used when \code{y} is \code{NULL}. Defaults to '#000000FF'.}

\item{locked.out.color}{\code{character} color to denote locked in planning units. Used when \code{y} is \code{NULL}. Defaults to '#D7D7D7FF'.}

\item{alpha}{\code{numeric} value to indicate how transparent the planning unit colors shoud be.}

\item{grayscale}{\code{logical} should the basemap be gray-scaled?}

\item{force.reset}{\code{logical} if basemap data has been cached, should it be re-downloaded?}

\item{i}{\code{NULL} to plot selection frequencies. \code{numeric} to plot the i'th solution, or 0 to plot the best solution. Only used when \code{y} is a \code{RapSolved} object. Defaults to \code{NULL}.}

\item{j}{\code{NULL} to plot selection frequencies. \code{numeric} to plot the i'th solution, or 0 to plot the best solution. Only used when \code{y} is a \code{RapSolved} object. Defaults to \code{j}.}

\item{x.locked.in.color}{\code{character} color to denote locked in planning units in \code{x}. Used when \code{y} is \code{RapSolved}. Defaults to '#000000FF'.}

\item{x.locked.out.color}{\code{character} color to denote locked out planning units in \code{x}. Used when \code{y} is \code{RapSolved}. Defaults to '#D7D7D7FF'.}

\item{y.locked.in.color}{\code{character} color to denote locked in planning units in \code{y}. Used when \code{y} is \code{RapSolved}. Defaults to '#FFFFFFFF'.}

\item{y.locked.out.color}{\code{character} color to denote locked out planning units in \code{y}. Used when \code{y} is \code{RapSolved}. Defaults to '#D7D7D7FF'.}
}
\description{
This function plots the solutions contained in \code{RapSolved} objects. It can be used to show a single solution, or the the selection frequencies of planning
units contained in a single \code{RapSolved} object. Additionally, two \code{RapSolved} objects can be supplied to plot the differences between them.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = RapSolved,y = numeric}: 

\item \code{x = RapSolved,y = missing}: 

\item \code{x = RapSolved,y = RapSolved}: 
}}
\examples{
# load data
data(sim_rs, cs_rs_amount, cs_rs_space)
## simulated species examples
# plot selection frequencies
plot(sim_rs)
# plot best solution
plot(sim_rs, 0)
# plot second solution
plot(sim_rs, 2)
# plot different between best and second solutions
plot(sim_rs, sim_rs, 0 ,2)
\dontrun{
## case-study species examples
# plot selection frequencies
plot(cs_rs_amount)
# plot best solution with basemap
plot(cs_rs_amount, 0, basemap='satellite')
plot(cs_rs_amount, 2, basemap='satellite', grayscale=TRUE)
# plot second solution
# plot different between best and second solutions
plot(cs_rs_amount, cs_rs_amount, 0, 2)
# plot difference in selection frequencies between two RapSolved objects
plot(cs_rs_amount, cs_rs_space)
# plot difference between best solutions in two RapSolved objects
plot(cs_rs_amount, cs_rs_space, 0, 0)
}
}
\seealso{
\code{\link{RapSolved}}.
}

