% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{plot_leading_edge}
\alias{plot_leading_edge}
\title{Plot expression of leading edge genes given enrichment result}
\usage{
plot_leading_edge(
  enrichment_result,
  pathway_name,
  dds,
  orthologs = human_mouse_orthologs,
  symbols = gene_symbols
)
}
\arguments{
\item{enrichment_result}{result of running run_fgsea}

\item{pathway_name}{pathway to plot leading genes}

\item{dds}{DESEQ2 dds object used for running fgsea}

\item{orthologs}{table containing mapping information between human and mouse gene symbols}

\item{symbols}{table containing mapping between mouse ensembl_ids and gene symbols}
}
\value{

}
\description{
Plot expression of leading edge genes given enrichment result
}
