% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{plotGeneSetEnrichment}
\alias{plotGeneSetEnrichment}
\title{Plot the results of GSEA}
\usage{
plotGeneSetEnrichment(
  enrichmentScores,
  plot.title = "Gene Set Enrichment Plot"
)
}
\arguments{
\item{enrichmentScores}{output of geneSetEnrichment}

\item{plot.title}{optional title for the output plot}
}
\value{

}
\description{
Plot the results of GSEA
}
