% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspSolved.R, R/generics.R
\docType{methods}
\name{plotRasp,RaspSolved,numeric-method}
\alias{plotRasp}
\alias{plotRasp,RaspSolved,RaspSolved-method}
\alias{plotRasp,RaspSolved,missing-method}
\alias{plotRasp,RaspSolved,numeric-method}
\title{Plot Rasp solutions}
\usage{
\S4method{plotRasp}{RaspSolved,numeric}(x, y, basemap = "none",
  colramp = "Greens", lockedincol = "#000000FF",
  lockedoutcol = "#D7D7D7FF", alpha = ifelse(basemap == "none", 1, 0.7),
  grayscale = FALSE, force_reset = FALSE)

\S4method{plotRasp}{RaspSolved,missing}(x, y, basemap = "none",
  colramp = "PuBu", lockedincol = "#000000FF", lockedoutcol = "#D7D7D7FF",
  alpha = ifelse(basemap == "none", 1, 0.7), grayscale = FALSE,
  force_reset = FALSE)

\S4method{plotRasp}{RaspSolved,RaspSolved}(x, y, i = NULL, j = i,
  basemap = "none", colramp = ifelse(is.null(i), "RdYlBu", "Accent"),
  xlockedincol = "#000000FF", xlockedoutcol = "#D7D7D7FF",
  ylockedincol = "#FFFFFFFF", ylockedoutcol = "#D7D7D7FF",
  alpha = ifelse(basemap == "none", 1, 0.7), grayscale = FALSE,
  force_reset = FALSE)

plotRasp(x, y, ...)
}
\arguments{
\item{x}{\code{RaspData}, \code{RaspUnsolved}, \code{RaspSolved} object.}

\item{y}{See below for details:
    \itemize{
    \item if \code{NULL}: function plots the selection frequency of planning units for all solutions (default).
    \item if \code{integer}: function plots the selection for the n'th solution, use 0 to plot the best solution.
    \item if \code{RaspSolved} plots the difference in solutions: \itemize{
        \item if \code{i} is \code{NULL}: differences in selection frequencies are plotted.
        \item if \code{i} is \code{integer} and \code{j} is \code{integer}, plots differences selection status for solution \code{i} in \code{x}, and solution \code{j} in \code{y}. Set \code{i} or \code{j} to 0 to refer to the best solution in \code{x} or \code{y} respectively.
        }
}}

\item{basemap}{\code{character} name of basemap to display. Valid names are "roadmap", "mobile", "satellite", "terrain", "hybrid", "mapmaker-roadmap", "mapmaker-hybrid".}

\item{colramp}{\code{character} name of colour palette (see \code{\link[RColorBrewer]{brewer.pal.info}}).}

\item{alpha}{\code{numeric} alpha value}

\item{grayscale}{\code{logical} should basemap be gray scale?}

\item{force_reset}{\code{logical} ignore data in cache? Setting this as ignore will make function slower but may avoid bugs in cache system.}

\item{xlockedincol}{\code{character} color to plot locked in planning units in object \code{x}.}

\item{xlockedoutcol}{\code{character} color to plot locked out planning units in object \code{x}.}

\item{ylockedincol}{\code{character} color to plot locked in planning units in object \code{y}.}

\item{ylockedoutcol}{\code{character} color to plot locked out planning units in object \code{y}.}
}
\description{
This function makes a geoplot displaying Rasp solutions.
}
\note{
This function will return an error if spatial polygons were not supplied during the construction of the Rasp object.
}
\seealso{
\code{\link{RaspResults-class}}, \code{\link{RaspSolved-class}}, \code{\link{basemap}} \code{\link{rasp}}, \code{\link{spplot}}
}

