% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RapData.R, R/RapSolved.R, R/RapUnsolved.R, R/generics.R
\name{prob.subset.RapData}
\alias{prob.subset}
\alias{prob.subset.RapData}
\alias{prob.subset.RapSolved}
\alias{prob.subset.RapUnsolved}
\title{Subset probabilities above a threshold}
\usage{
\method{prob.subset}{RapData}(x, species, threshold)

\method{prob.subset}{RapSolved}(x, species, threshold)

\method{prob.subset}{RapUnsolved}(x, species, threshold)

prob.subset(x, species, threshold)
}
\arguments{
\item{x}{\code{RapData}, \code{RapUnsolved}, or \code{RapSolved} object.}

\item{species}{\code{integer} vector to specify the index of the species to apply threshold to}

\item{threshold}{\code{numeric} probability to use a threshold.}
}
\value{
\code{RapData} or \code{RapUnsolved} object depending on input object.
}
\description{
This function subsets out probabilities assigned to planning units above a threshold. It effectively sets the probability
that species inhabit planning units to zero if they are below the threshold.
}
\examples{
# load data
data(sim_ru)
# generate new object with first 10 planning units
sim_ru2 <- prob.subset(sim_ru, 1:3, c(0.1,0.2,0.3))
}
\seealso{
\code{RapData}, \code{RapUnsolved}, \code{RapSolved}.
}

