% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RapData.R, R/RapSolved.R, R/RapUnsolved.R, R/generics.R
\name{pu.subset.RapData}
\alias{pu.subset}
\alias{pu.subset.RapData}
\alias{pu.subset.RapSolved}
\alias{pu.subset.RapUnsolved}
\title{Subset planning units}
\usage{
\method{pu.subset}{RapData}(x, pu)

\method{pu.subset}{RapSolved}(x, pu)

\method{pu.subset}{RapUnsolved}(x, pu)

pu.subset(x, pu)
}
\arguments{
\item{x}{\code{RapData}, \code{RapUnsolved}, or \code{RapSolved} object.}

\item{pu}{\code{integer} vector to specify the index of planning units to subset.}
}
\value{
\code{RapData} or \code{RapUnsolved} object depending on input object.
}
\description{
Subset planning units from a \code{RapData}, \code{RapUnsolved}, or \code{RapSolved} object.
}
\examples{
# load data
data(sim_ru)
# generate new object with first 10 planning units
sim_ru2 <- pu.subset(sim_ru, 1:10)
}
\seealso{
\code{RapData}, \code{RapUnsolved}, \code{RapSolved}.
}

