% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pus}
\alias{pus}
\title{Simulated planning units}
\format{\code{SpatialPolygonsDataFrame}}
\description{
A \code{SpatialPolygonsDataFrame} of containing simulated planning units. The columns in the data slot are 'id' (\code{integer}), 'cost' (\code{numeric}), and 'status' (\code{integer}).
}
\examples{
data(pus)
str(pus)
}
\keyword{datasets}

