% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{randomPoints}
\alias{randomPoints}
\title{Sample random points from a RasterLayer}
\usage{
randomPoints(mask, n, prob = FALSE)
}
\arguments{
\item{mask}{\code{RasterLayer} object}

\item{n}{\code{integer} number of points to sample}

\item{prob}{\code{logical} should the raster values be used as weights? Defaults to \code{FALSE}.}
}
\value{
code{matrix} with x-coordinates, y-coordinates, and cell values.
}
\description{
This function generates random points in a \code{RasterLayer} object.
}
\examples{
data(sim_spp)
# generate points
pts <- randomPoints(sim_spp[[1]])
# plot points
plot(sim_spp[[1]])
points(pts)
}
\seealso{
\code{\link[dismo]{randomPoints}}.
}

