% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{class}
\name{RapOpts-class}
\alias{RapOpts-class}
\alias{RapOpts}
\title{RapOpts class}
\description{
Object is either \code{\link[=RapReliableOpts]{RapReliableOpts()}} or
\code{\link[=RapUnreliableOpts]{RapUnreliableOpts()}}.
}
