% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapReliableOpts.R
\name{RapReliableOpts-class}
\alias{RapReliableOpts-class}
\title{RapReliableOpts: An S4 class to represent input parameters for the reliable
formulation of RAP.}
\description{
This class is used to store input parameters for the reliable formulation of
RAP.
}
\section{Slots}{

\describe{
\item{\code{BLM}}{\code{numeric} boundary length modifier. Defaults to 0.}

\item{\code{failure.multiplier}}{\code{numeric} multiplier for failure planning
unit. Defaults to 1.1.}

\item{\code{max.r.level}}{\code{numeric} maximum R failure level for approximation.
Defaults to 5L.}
}}

\seealso{
\code{\link{RapReliableOpts}}.
}
