% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RapResults.R
\name{RapResults}
\alias{RapResults}
\title{Create RapResults object}
\usage{
RapResults(summary, selections, amount.held, space.held, logging.file)
}
\arguments{
\item{summary}{\code{data.frame} with summary information on solutions.}

\item{selections}{\code{matrix} with binary selections.}

\item{amount.held}{\code{matrix} with the amount held for each species in each solution.}

\item{space.held}{\code{matrix} with the poportion of attribute space sampled for each species in each solution.}

\item{logging.file}{\code{character} Gurobi log files.}
}
\value{
\code{RapResults} object
}
\description{
This function creates a new \code{RapResults} object.
}
\note{
slot \code{best} is automatically determined based on data in \code{summary}.
}
\seealso{
\code{\link{RapResults-class}} \code{\link{read.RapResults}}
}

