% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapSolved.R
\docType{class}
\name{RapSolved-class}
\alias{RapSolved-class}
\title{RapSolved: An S4 class to represent RAP inputs and outputs}
\description{
This class is used to store RAP input and output data in addition to input parameters.
}
\section{Slots}{

\describe{
\item{\code{opts}}{\code{RapReliableOpts} or \code{RapUnreliableOpts} object used to store input parameters.}

\item{\code{solver}}{\code{GurobiOpts} or \code{ManualOpts} object used to store solver information/parameters.}

\item{\code{data}}{\code{RapData} object used to store input data.}

\item{\code{results}}{\code{RapResults} object used to store results.}
}}
\seealso{
\code{\link{RapReliableOpts-class}}, \code{\link{RapUnreliableOpts-class}}, \code{\link{RapData-class}}, \code{\link{RapResults-class}}.
}

