% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapSolved.R
\name{RapSolved}
\alias{RapSolved}
\title{Create new RapSolved object}
\usage{
RapSolved(unsolved, solver, results)
}
\arguments{
\item{unsolved}{\code{\link{RapUnsolved}} object.}

\item{solver}{\code{\link{GurobiOpts}} or \code{\link{ManualOpts}} object.}

\item{results}{\code{\link{RapResults}} object.}
}
\value{
\code{\link{RapSolved}} object.
}
\description{
This function creates a \code{\link{RapSolved}} object.
}
\seealso{
\code{\link{RapSolved-class}}, \code{\link{RapResults-class}},
  \code{link{solve}}.
}
