% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{raptr-deprecated}
\alias{raptr-deprecated}
\alias{SpatialPolygons2PolySet}
\title{Deprecation notice}
\usage{
SpatialPolygons2PolySet(...)
}
\arguments{
\item{...}{not used.}
}
\description{
The functions listed here are deprecated.
This means that they once existed in earlier versions of the
of the \pkg{raptr} package, but they have since been removed
entirely, replaced by other functions, or renamed as other functions
in newer versions.
To help make it easier to transition to new versions of the \pkg{raptr}
package, we have listed alternatives for deprecated the functions
(where applicable).
If a function is described as being renamed, then this means
that only the name of the function has changed
(i.e., the inputs, outputs, and underlying code remain the same).
}
\details{
The following functions have been deprecated:

\describe{

\item{\code{SpatialPolygons2PolySet()}}{renamed
as the \code{\link[=convert2PolySet]{convert2PolySet()}} function.}
}
}
\keyword{deprecated}
