% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapUnreliableOpts.R
\docType{class}
\name{RapUnreliableOpts-class}
\alias{RapUnreliableOpts-class}
\title{RapUnreliableOpts: An S4 class to represent parameters for the unreliable
RAP problem}
\description{
This class is used to store input parameters for the unreliable RAP problem
formulation.
}
\section{Slots}{

\describe{
\item{\code{BLM}}{\code{numeric} boundary length modifier. Defaults to 0.}
}}

