% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapUnreliableOpts.R
\name{RapUnreliableOpts}
\alias{RapUnreliableOpts}
\title{Create RapUnreliableOpts object}
\usage{
RapUnreliableOpts(BLM = 0)
}
\arguments{
\item{BLM}{\code{numeric} boundary length modifier. Defaults to 0.}
}
\value{
\code{\link{RapUnreliableOpts}} object
}
\description{
This function creates a new RapUnreliableOpts object.
}
\examples{
# create RapUnreliableOpts using defaults
RapUnreliableOpts(BLM = 0)

}
\seealso{
\code{\link{RapUnreliableOpts-class}}.
}
