% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspOpts.R
\docType{class}
\name{RaspOpts-class}
\alias{RaspOpts-class}
\title{RaspOpts: An S4 class to represent RASP input parameters}
\description{
This class is used to store RASP input parameters.
}
\section{Slots}{

\describe{
\item{\code{BLM}}{\code{numeric} boundary length modifier. Defaults to 0.}

\item{\code{NUMREPS}}{\code{integer} number of replicate runs. Defaults to 1L.}

\item{\code{NTHREADS}}{\code{integer} number of cores to use for processing. Defaults to 1L.}

\item{\code{GAP}}{\code{numeric} MIP gap (lp_solve parameter). Defaults to 0.05.}

\item{\code{PRESOLVE}}{\code{integer} code for level of computation in presolve (lp_solve parameter). Defaults to 2.}

\item{\code{TIMELIMIT}}{\code{integer} number of seconds to allow for solving. Defaults to NA_integer_, and so a time limit is not imposed.}

\item{\code{VERBOSITY}}{\code{integer} code for amount of output to display. Defaults to 1L.}
}}

