% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspOpts.R
\name{RaspOpts}
\alias{RaspOpts}
\title{Create RaspOpts object}
\usage{
RaspOpts(BLM = 0, FAILUREMULTIPLIER = 1.1, MAXRLEVEL = 5L, NUMREPS = 1L)
}
\arguments{
\item{BLM}{\code{numeric} boundary length modifier. Defaults to 0.}

\item{FAILUREMULTIPLIER}{\code{numeric} multiplier for failure planning unit. Defaults to 1.1.}

\item{MAXRLEVEL}{\code{numeric} maximum R failure level for approximation. Defaults to 5L.}

\item{NUMREPS}{\code{integer} number of replicate runs. Defaults to 1L.}
}
\value{
\code{MarxanOpts} object
}
\description{
This function creates a new RaspOpts object.
}
\examples{
# create RaspOpts using defaults
x <- RaspOpts(BLM=0, FAILUREMULTIPLIER=1.1, MAXRLEVEL=5L, NUMREPS=1L)
print(x)
}
\seealso{
\code{\link{RaspOpts-class}}.
}

