% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/raspr.R
\docType{package}
\name{raspr}
\alias{raspr}
\alias{raspr-package}
\title{raspr: identify prioritisations using the Representative and Adequate Sample Problem (RASP)}
\description{
Biodiversity is in crisis. The overarching aim of conservation
is to preserve biodiversity patterns and processes. To this end, protected
areas are established to buffer species and preserve biodiversity processes.
But resources are limited and so protected areas must be cost-effective.
This package contains functions to generate plans for protected areas--
prioritisations--using the Representative and Adequate Sample Problem (RASP).
Reserve selection problems are expressed as a mixed linear integer problem
(MILP) and solved using exact-algorithm methods. Prioritisations can be obtained
using Gurobi (obtained from <http://www.gurobi.com/>).
}
\details{
The main classes used in this package are used to store input data and prioritisations:
\itemize{
  \item \code{\link{GurobiOpts-class}}: parameters for solving optimisation problems using Gurobi.
  \item \code{\link{RaspReliableOpts-class}}: parameters for the reliable formulation of RASP.
  \item \code{\link{RaspUnreliableOpts-class}}: parameters for the unreliable formulation of RASP.
  \item \code{\link{RaspData-class}}: planning unit, species data, and demand points for RASP.
  \item \code{\link{RaspUnsolved-class}}: contains all the data and input parameters required to generate prioritisations using RASP. This class contains a \code{\link{GurobiOpts-class}} object, a \code{\link{RaspReliableOpts-class}} or \code{\link{RaspUnreliableOpts-class}} object, and a \code{\link{RaspData-class}} object.
  \item \code{\link{RaspResults-class}}: prioritisations and summary statistics on their performance.
  \item \code{\link{RaspSolved-class}}: contains all the input data, parameters and output data. This class contains all the objects in a \code{\link{RaspUnsolved}} object and also a \code{\link{RaspResults-class}} object.
}

Type \code{vignette('raspr')} for a tutorial on how to use this package.
}

