% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspReliableOpts.R
\docType{class}
\name{RaspReliableOpts-class}
\alias{RaspReliableOpts-class}
\title{RaspReliableOpts: An S4 class to represent input parameters for the reliable formulation of RASP.}
\description{
This class is used to store input parameters for the reliable formulation of RASP.
}
\section{Slots}{

\describe{
\item{\code{BLM}}{\code{numeric} boundary length modifier. Defaults to 0.}

\item{\code{FAILUREMULTIPLIER}}{\code{numeric} multiplier for failure planning unit. Defaults to 1.1.}

\item{\code{MAXRLEVEL}}{\code{numeric} maximum R failure level for approximation. Defaults to 5L.}

\item{\code{NUMREPS}}{\code{integer} number of replicate runs. Defaults to 1L.}
}}

