% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R
\name{RaspResults}
\alias{RaspResults}
\title{Create RaspResults object}
\usage{
RaspResults(summary, selections, amount.held, space.held, model.file, log.file,
  solution.file)
}
\arguments{
\item{summary}{\code{data.frame} with summary information on solutions.}

\item{selections}{\code{matrix} with binary selections.}

\item{amount.held}{\code{matrix} with the amount held for each species in each solution.}

\item{space.held}{\code{matrix} with the poportion of attribute space sampled for each species in each solution.}

\item{model.file}{\code{character} with Gurobi model file.}

\item{log.file}{\code{character} with Gurobi log file.}

\item{solution.file}{\code{character} with Gurobi solution file.}
}
\value{
\code{RaspResults} object
}
\description{
This function creates a new \code{RaspResults} object.
}
\note{
slot \code{best} is automatically determined based on data in \code{summary}.
}
\seealso{
\code{\link{RaspResults-class}} \code{\link{read.RaspResults}}
}

