% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspSolved.R
\name{RaspSolved}
\alias{RaspSolved}
\title{Create new RaspSolved object}
\usage{
RaspSolved(unsolved, results)
}
\arguments{
\item{unsolved}{\code{RaspUnsolved} object.}

\item{results}{\code{RaspResults} object.}
}
\value{
\code{RaspSolved} object.
}
\description{
This function creates a \code{RaspSolved} object using a \code{RaspUnsolved} object and a \code{RaspResults} object.
}
\examples{
\dontrun{
data(sim_pus, sim_spp)
# create inputs for a RaspUnsolved  object
go <- GurobiOpts(MIPGap=0.9)
rd <- make.RaspData(
	sim_pus[1:10,],
	sim_spp[[1]],
	NULL,
	include.geographic.space=TRUE,
	n.demand.points=5L
)
ro <- RaspOpts(NUMREPS=1L)
# create RaspUnsolved object
ru <- RaspUnsolved(ro1, go, rd)
# solve it to return a RaspSolved object
rs1 <- (ru)
# methods for RaspSolved
print(rs)
summary(rs)
selections(rs)
score(rs)
amount.held(rs)
space.held(rs)
logging.file(rs)
plot(rs)
}
}
\seealso{
\code{\link{RaspSolved-class}}, \code{\link{RaspResults-class}}.
}

