% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspUnreliableOpts.R
\docType{class}
\name{RaspUnreliableOpts-class}
\alias{RaspUnreliableOpts-class}
\title{RaspUnreliableOpts: An S4 class to represent parameters for the unreliable RASP problem}
\description{
This class is used to store input parameters for the unreliable RASP problem formulation.
}
\section{Slots}{

\describe{
\item{\code{BLM}}{\code{numeric} boundary length modifier. Defaults to 0.}
}}

