% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspUnreliableOpts.R
\name{RaspUnreliableOpts}
\alias{RaspUnreliableOpts}
\title{Create RaspUnreliableOpts object}
\usage{
RaspUnreliableOpts(BLM = 0)
}
\arguments{
\item{BLM}{\code{numeric} boundary length modifier. Defaults to 0.}
}
\value{
\code{RaspUnreliableOpts} object
}
\description{
This function creates a new RaspUnreliableOpts object.
}
\examples{
# create RaspUnreliableOpts using defaults
RaspUnreliableOpts(BLM=0)
}
\seealso{
\code{\link{RaspUnreliableOpts-class}}.
}

