% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspUnsolved.R
\docType{class}
\name{RaspUnsolved-class}
\alias{RaspUnsolved-class}
\title{RaspUnsolved: An S4 class to represent RASP inputs}
\description{
This class is used to store RASP input data and input parameters.
}
\section{Slots}{

\describe{
\item{\code{opts}}{\code{RaspReliableOpts} or \code{RaspUnreliableOpts} object used to store input parameters.}

\item{\code{solver}}{\code{GurobiOpts} object used to store Gurobi parameters.}

\item{\code{data}}{\code{RaspData} object used to store input data.}
}}
\seealso{
\code{\link{RaspReliableOpts-class}}, \code{\link{RaspUnreliableOpts-class}},  \code{\link{GurobiOpts-class}}, \code{\link{RaspData-class}}.
}

