% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspUnsolved.R
\name{RaspUnsolved}
\alias{RaspUnsolved}
\title{Create a new RaspUnsolved object}
\usage{
RaspUnsolved(opts, gurobi, data)
}
\arguments{
\item{opts}{\code{RaspOpts} object.}

\item{gurobi}{\code{GurobiOpts} object used to store Gurobi parameters.}

\item{data}{\code{RaspData} object.}
}
\value{
\code{RaspUnsolved} object.
}
\description{
This function creates a \code{RaspUnsolved} object using \code{RaspOpts} and \code{RaspData} objects.
}
\seealso{
\code{\link{RaspOpts-class}},  \code{\link{GurobiOpts-class}}, \code{\link{RaspData-class}}.
}

