% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raptr-internal.R
\name{read.RapResults}
\alias{read.RapResults}
\title{Read RAP results}
\usage{
read.RapResults(opts, data, model.list, logging.file, solution.list,
  verbose = FALSE)
}
\arguments{
\item{opts}{\code{RapReliableOpts} or \code{RapUnreliableOpts} object}

\item{data}{\code{RapData} object}

\item{model.list}{\code{list} object containing Gurobi model data}

\item{logging.file}{\code{character} Gurobi log files.}

\item{solution.list}{\code{list} object containing Gurobi solution data.}

\item{verbose}{\code{logical} print progress messages? Defaults to
\code{FALSE}.}
}
\value{
\code{RapResults} object
}
\description{
This function reads files output from Gurobi and returns a \code{RapResults}
object.
}
\seealso{
\code{\link{RapReliableOpts}}, \code{\link{RapUnreliableOpts}},
  \code{\link{RapData}}, \code{\link{RapResults}}.
}
\keyword{internal}
