% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/raspr-internal.R
\name{read.RaspResults}
\alias{read.RaspResults}
\title{Read RASP results}
\usage{
read.RaspResults(opts, data, model.list, logging.file, solution.list)
}
\arguments{
\item{opts}{\code{RaspOpts} object}

\item{data}{\code{RaspData} object}

\item{model.list}{\code{list} object containing Gurobi model data}

\item{logging.file}{\code{character} Gurobi log files.}

\item{solution.list}{\code{list} object containing Gurobi solution data.}
}
\value{
\code{RaspResults} object
}
\description{
This function reads files output from Gurobi and returns a \code{RaspResults} object.
}
\seealso{
\code{\link{RaspData}}, \code{\link{RaspData-class}}, \code{\link{RaspResults-class}}, \code{\link{RaspResults}}.
}
\keyword{internal}

