% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{read_enrichR}
\alias{read_enrichR}
\title{Import a collection of gene sets downloaded from enrichR}
\usage{
read_enrichR(file_path)
}
\arguments{
\item{file_path}{path to file containing gene sets}
}
\value{

}
\description{
Import a collection of gene sets downloaded from enrichR
}
