% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{run_fgsea}
\alias{run_fgsea}
\title{Run GSEA given differential expression analysis and list of gene sets}
\usage{
run_fgsea(
  resObj,
  gene_sets,
  orthologs = human_mouse_orthologs,
  symbols = gene_symbols
)
}
\arguments{
\item{resObj}{ouput of running results on a DESEQ2 dds object}

\item{gene_sets}{list of gene sets to run enrichment. See read_enrichR}

\item{orthologs}{table containing mapping information between human and mouse gene symbols}

\item{symbols}{table containing mapping between mouse ensembl_ids and gene symbols}
}
\value{

}
\description{
Run GSEA given differential expression analysis and list of gene sets
}
