% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R, R/RaspSolved.R, R/generics.R
\name{score.RaspResults}
\alias{score}
\alias{score.RaspResults}
\alias{score.RaspSolved}
\title{Extract solution score}
\usage{
\method{score}{RaspResults}(x, y = 0)

\method{score}{RaspSolved}(x, y = 0)

score(x, ...)
}
\arguments{
\item{x}{\code{RaspResults} or \code{RaspSolved} object.}

\item{y}{"NULL" to return all scores, "integer" 0 to return score for best solution, "integer" value greater than 0 for \code{y}'th solution score.}
}
\value{
"matrix" or "numeric" vector with solution score(s) depending on arguments.
}
\description{
Extract solution score from \code{RaspResults} or \code{RaspSolved} object.
}
\seealso{
\code{\link{RaspResults-class}}, \code{\link{RaspSolved-class}}, \code{\link{rasp}}
}

