% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R, R/RaspSolved.R, R/generics.R
\name{selections.RaspResults}
\alias{selections}
\alias{selections.RaspResults}
\alias{selections.RaspSolved}
\title{Extract solution selections}
\usage{
\method{selections}{RaspResults}(x, y = 0)

\method{selections}{RaspSolved}(x, y = 0)

selections(x, y)
}
\arguments{
\item{x}{\code{RaspResults} or \code{RaspSolved} object.}

\item{y}{\code{NULL} to return all values, \code{integer} 0 to return values for best solution, \code{integer} value greater than 0 for \code{y}'th solution value.}

\item{...}{not used.}
}
\value{
\code{matrix} or \code{numeric} vector depending on arguments.
}
\description{
Extract selections for a given solution from a \code{RaspResults} or \code{RaspSolved} object.
}
\examples{
load(sim_rs)
# selections for best solution
selections(sim_rs, 0)
# selections for second solution
selections(sim_rs, 2)
# selections for each solution
selections(sim_rs)
}
\seealso{
\code{RaspResults}, \code{RaspSolved}.
}

