% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GurobiOpts.R, R/RaspData.R, R/RaspOpts.R, R/RaspResults.R, R/RaspSolved.R, R/RaspUnsolved.R, R/generics.R
\docType{methods}
\name{show,GurobiOpts-method}
\alias{show}
\alias{show,GurobiOpts-method}
\alias{show,RaspData-method}
\alias{show,RaspOpts-method}
\alias{show,RaspResults-method}
\alias{show,RaspSolved-method}
\alias{show,RaspUnsolved-method}
\title{Show objects}
\usage{
\S4method{show}{GurobiOpts}(object)

\S4method{show}{RaspData}(object)

\S4method{show}{RaspOpts}(object)

\S4method{show}{RaspResults}(object)

\S4method{show}{RaspSolved}(object)

\S4method{show}{RaspUnsolved}(object)
}
\arguments{
\item{object}{\code{GurobiOpts}, \code{RaspOpts}, \code{RaspData}, \code{RaspUnsolved}, \code{RaspResults}, or \code{RaspSolved} object.}
}
\description{
Shows objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{GurobiOpts}: 

\item \code{RaspData}: 

\item \code{RaspOpts}: 

\item \code{RaspResults}: 

\item \code{RaspSolved}: 

\item \code{RaspUnsolved}: 
}}

