% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimplePoints.R
\docType{class}
\name{SimplePoints-class}
\alias{SimplePoints-class}
\title{SimplePoints: An S4 class to represent points}
\description{
This class is used to store \code{d}-dimensional point coordinates.
}
\section{Slots}{

\describe{
\item{\code{coords}}{\code{matrix} coordinates for each point.}
}}
\seealso{
\code{\link{DemandPoints}}, \code{\link{AttributeSpace}}.
}

