% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimplePoints.R
\name{SimplePoints}
\alias{SimplePoints}
\title{Create new SimplePoints object}
\usage{
SimplePoints(coords)
}
\description{
This function creates a new \code{SimplePoints} object.
}
\section{Slots}{

\describe{
\item{\code{coords}}{\code{matrix} coordinates for each point.}
}}
\seealso{
\code{\link{DemandPoints-class}}, \code{\link{AttributeSpace-class}}.
}

