% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.pus.R
\name{simulate.pus}
\alias{simulate.pus}
\title{Simulate planning units}
\usage{
\method{simulate}{pus}(n, xmn = -sqrt(n)/2, xmx = sqrt(n)/2,
  ymn = -sqrt(n)/2, ymx = sqrt(n)/2, ...)
}
\arguments{
\item{n}{\code{integer} number of planning units. \code{sqrt(n)} must yield a valid number.}

\item{xmn}{\code{numeric} value for minimim x-coordinate.}

\item{xmx}{\code{numeric} value for maximum x-coordinate.}

\item{ymn}{\code{numeric} value for minimim y-coordinate.}

\item{ymx}{\code{numeric} value for maximum y-coordinate.}
}
\value{
\code{SpatialPolygons} with planning units.
}
\description{
This function simulates planning units for RASP.
}
\details{
Square planning units are generated in the shape of a square. Default coordinate arguments are such that the planning units will be centered at origin.
The data slot contains an 'id' (\code{integer}), cost (\code{numeric}), 'status' (\code{integer}), and area (\code{numeric}).
}

