% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/simulate.space.R
\name{simulate.space}
\alias{simulate.space}
\alias{simulate.space.RasterLayer}
\alias{simulate.space.SpatialPolygons}
\title{Simulate attribute space data for RASP}
\usage{
simulate.space(x, ...)

\method{simulate}{space.RasterLayer}(x, d = 2, model = RMgauss(), ...)

\method{simulate}{space.SpatialPolygons}(x, res, d = 2, model = RMgauss(),
  ...)
}
\arguments{
\item{x}{\code{RasterLayer} or \code{SpatialPolygons} object delineate the spatial extent to delineate study area.}

\item{...}{parameters passed to \code{\link[RandomFields]{RandomFields}}.}

\item{d}{\code{integer} number of dimensions. Defaults to 2.}

\item{model}{\code{RMmodel} model to simulate species distributions with. Defaults \code{\link[RandomFields]{RPgauss}}.}

\item{res}{\code{numeric} resolution to simulate distributions. Only needed when \code{SpatialPolygons} supplied.}
}
\value{
\code{RasterStack} with layers for each dimension of the space.
}
\description{
This function simulates attribute space data for RASP.
}
\details{
Distributions are simulated by passing \code{model} to \code{\link[RandomFields]{RFsimulate}}.
}
\seealso{
\code{\link[RandomFields]{RFsimulate}}
}

