% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/simulate.species.R
\name{simulate.species}
\alias{simulate.species}
\alias{simulate.species.RasterLayer}
\alias{simulate.species.SpatialPolygons}
\title{Simulate species distribution data for RASP}
\usage{
simulate.species(x, ...)

\method{simulate}{species.RasterLayer}(x, n = 1, model = RPgauss(), ...)

\method{simulate}{species.SpatialPolygons}(x, res, n = 1, model = RPgauss(),
  ...)
}
\arguments{
\item{x}{\code{RasterLayer} or \code{SpatialPolygons} object delineate the spatial extent to delineate study area.}

\item{...}{parameters passed to \code{\link[RandomFields]{RandomFields}}.}

\item{n}{\code{integer} number of species. Defaults to 1.}

\item{model}{\code{RMmodel} model to simulate species distributions with. Defaults \code{\link[RandomFields]{RPgauss}}.}

\item{res}{\code{numeric} resolution to simulate distributions. Only needed when \code{SpatialPolygons} supplied.}
}
\value{
\code{RasterStack} with layers for each species.
}
\description{
This function simulates species distributions for RASP.
}
\details{
Distributions are simulated by passing \code{model} to \code{\link[RandomFields]{RFsimulate}} and converting to logistic values using \code{\link[boot]{inv.logit}}.
}
\seealso{
\code{\link[RandomFields]{RFsimulate}}
}

