% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small.pus}
\alias{small.pus}
\title{Simulated planning units (small dataset)}
\format{\code{SpatialPolygonsDataFrame}}
\description{
A \code{SpatialPolygonsDataFrame} of containing simulated planning units. The columns in the data slot are 'cost' (\code{numeric}), and 'status' (\code{integer}).
}
\examples{
data(small.pus)
str(small.pus)
}
\keyword{datasets}

