% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small.species}
\alias{small.species}
\title{Simulated species distributions (small dataset)}
\format{\code{RasterStack}}
\description{
A \code{RasterStack} of containing simulated distributions for one species. Each pixel denotes the probability of occupancy.
}
\examples{
data(small.species)
str(small.species)
}
\keyword{datasets}

