% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R, R/RaspSolved.R, R/generics.R
\name{solution.file.RaspResults}
\alias{solution.file}
\alias{solution.file.RaspResults}
\alias{solution.file.RaspSolved}
\title{Solution file}
\usage{
\method{solution.file}{RaspResults}(x)

\method{solution.file}{RaspSolved}(x)

solution.file(x)
}
\arguments{
\item{x}{\code{RaspResults} or \code{RaspSolved} object.}
}
\value{
\code{character} object.
}
\description{
This function returns the Gurobi solution file (*.sol) associated with solving RASP.
}
\seealso{
\code{\link{RaspResults-class}}, \code{\link{RaspSolved}}, \code{\link{rasp}}.
}

