% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R, R/RaspSolved.R, R/RaspUnsolved.R, R/generics.R
\name{space.plot.RaspData}
\alias{space.plot}
\alias{space.plot.RaspData}
\alias{space.plot.RaspSolved}
\alias{space.plot.RaspUnsolved}
\title{Plot space}
\usage{
\method{space.plot}{RaspData}(x, species, space = 1,
  pu.color.palette = "RdYlGn", locked.in.color = "#000000FF",
  locked.out.color = "#D7D7D7FF")

\method{space.plot}{RaspSolved}(x, species, space = 1, y = 0,
  pu.color.palette = "RdYlGn", locked.in.color = "#000000FF",
  locked.out.color = "#D7D7D7FF")

\method{space.plot}{RaspUnsolved}(x, y, space = 1,
  pu.color.palette = "RdYlGn", locked.in.color = "#000000FF",
  locked.out.color = "#D7D7D7FF")

space.plot(x, species, space, ...)
}
\arguments{
\item{x}{\code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.}

\item{species}{\code{character} name of species, or \code{integer} index for species.}

\item{space}{\code{integer} index of attribute space.}

\item{pu.color.palette}{\code{character} name of color palette to use for planning units (see \code{\link[RColorBrewer]{brewer.pal}}). Defaults to 'RdYlGn'.}

\item{locked.in.color}{\code{character} color to denote locked in planning units. Used when \code{y} is \code{NULL}. Defaults to '#000000FF'.}

\item{locked.out.color}{\code{character} color to denote locked in planning units. Used when \code{y} is \code{NULL}. Defaults to '#D7D7D7FF'.}

\item{y}{\code{NULL} \code{integer} 0 to return values for best solution, \code{integer} value greater than 0 for \code{y}'th solution value.}

\item{...}{not used.}
}
\description{
This function plots the distribution of planning units and the distribution of demand points for a particular species in an attribute space.
Note that this function only works for attribute spaces with one, two, or three dimensions.
}
\examples{
# load RaspSolved objects
data(sim_rs)
# plot distribution of first species in first attribute space
space.plot(sim_rs, 1, 1)
}

