% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{space}
\alias{space}
\title{Simulated space data}
\format{\code{RasterStack}}
\description{
A \code{RasterStack} of containing simulated projections for space. This space has two dimensions.
}
\examples{
data(space)
str(space)
}
\keyword{datasets}

