% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R, R/RaspSolved.R, R/RaspUnsolved.R, R/generics.R
\name{space.target.RaspData}
\alias{space.target}
\alias{space.target.RaspData}
\alias{space.target.RaspSolved}
\alias{space.target.RaspUnsolved}
\title{Extract attribute space targets}
\usage{
\method{space.target}{RaspData}(x)

\method{space.target}{RaspSolved}(x)

\method{space.target}{RaspUnsolved}(x)

space.target(x)
}
\arguments{
\item{x}{\code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.}
}
\value{
code{numeric} vector.
}
\description{
This function returns the attribute space targets for each species.
}
\examples{
# load data
load(sim_rs)
# extract space targets
space.target(sim_rs)
}
\seealso{
\code{RaspData}, \code{RaspResults}, \code{RaspSolved}.
}

