% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{Simulated species distributions}
\format{\code{RasterStack}}
\description{
A \code{RasterStack} of containing simulated distributions for five species. Each pixel denotes the probability of occupancy.
}
\examples{
data(species)
str(species)
}
\keyword{datasets}

