% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_string.R
\name{split_string}
\alias{split_string}
\title{Split a string at every position of a delimiter, and returns the subset at a specified index.}
\usage{
split_string(string_vector, return_position = 1, sep_char = "_", fixed = TRUE)
}
\arguments{
\item{string_vector}{the string to split}

\item{return_position}{the index of the split string to return}

\item{sep_char}{the delimiter to use to split the string}

\item{fixed}{logical passed to strsplit. If TRUE match split exactly, otherwise use regular expressions.}
}
\description{
Split a string at every position of a delimiter, and returns the subset at a specified index.
}
