% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R, R/RaspSolved.R, R/RaspUnsolved.R, R/generics.R
\name{spp.subset.RaspData}
\alias{spp.subset}
\alias{spp.subset.RaspData}
\alias{spp.subset.RaspSolved}
\alias{spp.subset.RaspUnsolved}
\title{Subset species}
\usage{
\method{spp.subset}{RaspData}(x, species)

\method{spp.subset}{RaspSolved}(x, species)

\method{spp.subset}{RaspUnsolved}(x, species)

spp.subset(x, species)
}
\arguments{
\item{x}{\code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.}

\item{species}{\code{integer}, or \code{character} vectors to specify the index or species names to subset.}
}
\value{
\code{RaspData} or \code{RaspUnsolved} object depending on input object.
}
\description{
Subset species from a \code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.
}
\examples{
# load data
data(sim_ru)
# generate new object with only species 1
sim_ru2 <- spp.subset(sim_ru, 1)
}
\seealso{
\code{RaspData}, \code{RaspUnsolved}, \code{RaspSolved}.
}

