% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{summary}
\alias{summary}
\title{Summary of solutions}
\arguments{
\item{object}{\code{RapResults}, or \code{RapSolved} object.}

\item{...}{not used.}
}
\value{
\code{data.frame}
}
\description{
Extracts summary of solutions in a \code{RapResults} or \code{RapSolved} object.
}
\examples{
# load data
data(sim_rs)
# show summary
summary(sim_rs)
}
\seealso{
\code{\link{RapResults}}, \code{\link{RapSolved}}.
}

