% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_theme.R
\name{theme_custom}
\alias{theme_custom}
\title{theme_custom}
\usage{
theme_custom(
  base_size = 16,
  x_label_size = NA,
  y_label_size = NA,
  x_tick_label_size = NA,
  y_tick_label_size = NA,
  y_label_angle = 90,
  title_size = NA,
  title_just = 0.5,
  legend_position = "top"
)
}
\arguments{
\item{y_label_angle}{Angle to rotate y axis label.}

\item{legend_position}{}
}
\value{

}
\description{
theme_custom
}
