% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_theme.R
\name{theme_rapt}
\alias{theme_rapt}
\title{theme_rapt}
\usage{
theme_rapt(
  base_size = 16,
  x_label_size = NA,
  y_label_size = NA,
  x_tick_label_size = NA,
  y_tick_label_size = NA,
  y_label_angle = 90,
  title_size = NA
)
}
\arguments{
\item{y_label_angle}{Angle to rotate y axis label.}

\item{title_just}{}
}
\value{

}
\description{
theme_rapt
}
