% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_tibble.R
\name{transpose_tibble}
\alias{transpose_tibble}
\title{transpose_tibble}
\usage{
transpose_tibble(tbl, row.names = NA, col_label = "colnames")
}
\arguments{
\item{tbl}{the tibble to transpose}

\item{row.names}{name of the variable that should be used as new column names}

\item{col_label}{string that will be used as variable name for the new variable containing old column names}
}
\value{

}
\description{
transpose_tibble
}
