% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_z.R
\name{trunc_range}
\alias{trunc_range}
\title{Truncates a values to lower and/or upper bounds}
\usage{
trunc_range(
  x,
  low = NA,
  high = NA,
  filter = FALSE,
  low.val = NA,
  high.val = NA
)
}
\arguments{
\item{x}{input vector/matrix of values}

\item{low}{lower limit to truncate to}

\item{high}{upper limit to truncate to}

\item{filter}{if code{TRUE} filters out truncated values, otherwise replace them}

\item{low.val}{an alternate value to replace the lower limit with, rather than the lower limit itself}

\item{high.val}{an alternate value to replace the upper limit with, rather than the upper limit itself}
}
\description{
Truncates a values to lower and/or upper bounds
}
\examples{
x <- rnorm(500, mean=0, sd=5)
low.val/high.val is useful for differentiating truncated values on a plot
x1 <- trunc_range(x, low=-5, high=5, low.val=-6, high.val=6)
plot(x1 ~ x)
}
