% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R, R/RaspSolved.R, R/generics.R
\name{amount.held.RaspResults}
\alias{amount.held}
\alias{amount.held.RaspResults}
\alias{amount.held.RaspSolved}
\title{Extract amount held for a solution}
\usage{
\method{amount.held}{RaspResults}(x, y = 0)

\method{amount.held}{RaspSolved}(x, y = 0)

amount.held(x, y)
}
\arguments{
\item{x}{\code{RaspResults} or \code{RaspSolved} object.}

\item{y}{\code{NULL} to return all values, \code{integer} 0 to return values for best solution, \code{integer} value greater than 0 for \code{y}'th solution value.}
}
\value{
\code{matrix} or \code{numeric} vector depending on arguments.
}
\description{
This function returns the amount held for each species in a solution.
}
\seealso{
\code{\link{RaspResults-class}}, \code{\link{RaspSolved}}, \code{\link{rasp}}.
}

