% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RapSolved.R, R/generics.R
\name{amount.held.RapSolved}
\alias{amount.held}
\alias{amount.held.RapSolved}
\title{Extract amount held for a solution}
\usage{
\method{amount.held}{RapSolved}(x, y = 0, species = NULL)

amount.held(x, y, species)
}
\arguments{
\item{x}{\code{RapResults} or \code{RapSolved} object.}

\item{y}{\code{NULL} to return all values, \code{integer} 0 to return values for best solution, \code{integer} value greater than 0 for \code{y}'th solution value.}

\item{species}{\code{NULL} for all species or \code{integer} indicating species.}
}
\value{
\code{matrix} or \code{numeric} vector depending on arguments.
}
\description{
This function returns the amount held for each species in a solution.
}
\examples{
data(sim_rs)
# amount held (\\\%) for each species in best solution
amount.held(sim_rs, 0)
# amount held (\\\%) for each species in second solution
amount.held(sim_rs, 2)
# amount held (\\\%) for each species in each solution
amount.held(sim_rs, NULL)
}
\seealso{
\code{RapResults}, \code{RapSolved}.
}

