% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R, R/RaspSolved.R, R/RaspUnsolved.R, R/generics.R
\name{amount.target.RaspData}
\alias{amount.target}
\alias{amount.target.RaspData}
\alias{amount.target.RaspSolved}
\alias{amount.target.RaspUnsolved}
\title{Extract amount targets}
\usage{
\method{amount.target}{RaspData}(x)

\method{amount.target}{RaspSolved}(x)

\method{amount.target}{RaspUnsolved}(x)

amount.target(x)
}
\arguments{
\item{x}{\code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.}
}
\value{
code{numeric} vector.
}
\description{
This function returns the amount targets for each species.
}
\examples{
# load data
load(sim_rs)
# extract amount targets
amount.target(sim_rs)
}
\seealso{
\code{RaspData}, \code{RaspResults}, \code{RaspSolved}.
}

