% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GurobiOpts.R, R/generics.R
\name{as.list.GurobiOpts}
\alias{as.list}
\alias{as.list.GurobiOpts}
\title{Convert object to list.}
\usage{
\method{as.list}{GurobiOpts}(x, ...)
}
\arguments{
\item{x}{\code{GurobiOpts} object.}

\item{...}{not used.}
}
\description{
Convert objects to list.
}
\seealso{
\code{GurobiOpts}.
}

