% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/GurobiOpts.R
\name{as.list}
\alias{as.list}
\alias{as.list.GurobiOpts}
\title{Convert object to list.}
\usage{
\method{as.list}{GurobiOpts}(x, ...)
}
\arguments{
\item{x}{\code{\link{GurobiOpts}} object.}

\item{...}{not used.}
}
\value{
\code{list}
}
\description{
Convert \code{\link{GurobiOpts}} object to list.
}
\note{
This function will not include the \code{NumberSolutions} slot, the \code{MultipleSolutionsMethod} slot, or the \code{TimeLimit} slot if it is not finite.
}
\examples{
# make GuboriOpts object
x <- GurobiOpts()
# convert to list
as.list(x)
}
\seealso{
\code{GurobiOpts}.
}
