% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttributeSpace.R
\name{AttributeSpace-class}
\alias{AttributeSpace-class}
\title{AttributeSpace: An S4 class to represent an attribute space.}
\description{
This class is used to store planning unit points and demand points for a
single species in an attribute space.
}
\section{Slots}{

\describe{
\item{\code{planning.unit.points}}{\code{\link{PlanningUnitPoints}} for planning
unit in the space.}

\item{\code{demand.points}}{\code{\link{DemandPoints}} object for the space.}

\item{\code{species}}{\code{integer} species id to indicate which species the space
is associated with.}
}}

\seealso{
\code{\link{DemandPoints-class}},
  \code{\link{PlanningUnitPoints-class}}.
}
