% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R, R/RaspSolved.R, R/generics.R
\name{basemap.RaspData}
\alias{basemap}
\alias{basemap.RaspData}
\alias{basemap.RaspSolved}
\title{Basemap}
\usage{
\method{basemap}{RaspData}(x, basemap = "hybrid", grayscale = FALSE,
  force.reset = FALSE)

\method{basemap}{RaspSolved}(x, basemap = "none", grayscale = FALSE,
  force.reset = FALSE)

basemap(x, basemap = "hybrid", grayscale = FALSE, force.reset = FALSE)
}
\arguments{
\item{x}{\code{RaspData}, \code{RaspUnsolved}, \code{RaspSolved} object.}

\item{basemap}{\code{character} type of base map to display. Valid names are "roadmap", "mobile", "satellite", "terrain", "hybrid", "mapmaker-roadmap", "mapmaker-hybrid".}

\item{grayscale}{\code{logical} should base map be gray scale?}

\item{force.reset}{\code{logical} ignore data in cache? Setting this as ignore will make function slower but may avoid bugs in cache system.}
}
\value{
\code{list} with google map data.
}
\description{
This function retrieves google map data for planning units. The google map data is cached to provide fast plotting capabilities.
}
\seealso{
\code{\link[RgoogleMaps]{GetMap.bbox}}, \code{\link{plot}}.
}

