% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{blank.raster}
\alias{blank.raster}
\title{Blank raster}
\usage{
blank.raster(x, res)
}
\arguments{
\item{x}{\code{\link[sf:sf]{sf::st_sf()}} object.}

\item{res}{\code{numeric} \code{vector} specifying resolution of the output raster
in the x and y dimensions. If \code{vector} is of length one, then the
pixels are assumed to be square.}
}
\description{
This functions creates a blank \code{SpatRaster} based on the spatial extent of a
\code{sf} object.
}
\examples{
\dontrun{
# make sf data
polys <- sim.pus(225L)

# make raster from sf
blank.raster(polys, 1)
}
}
