% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R, R/RaspResults.R, R/raspr-internal.R
\docType{methods}
\name{cache,RaspData,character,ANY-method}
\alias{cache}
\alias{cache,RaspData,character,ANY-method}
\alias{cache,RaspData,character,missing-method}
\alias{cache,RaspResults,character,ANY-method}
\alias{cache,RaspResults,character,missing-method}
\title{Get and set cache Methods}
\usage{
\S4method{cache}{RaspData,character,ANY}(x, name, y)

\S4method{cache}{RaspData,character,missing}(x, name, y)

\S4method{cache}{RaspResults,character,ANY}(x, name, y)

\S4method{cache}{RaspResults,character,missing}(x, name, y)

cache(x, name, y)
}
\arguments{
\item{x}{\code{RaspData} or \code{RaspResults} object}

\item{name}{\code{character} hash.}

\item{y}{if \code{ANY} this object gets cached with name, else if \code{missing} the object hashed at name gets returned.}
}
\value{
\code{ANY} or \code{NULL} depends on \code{y} argument.
}
\description{
Getter and setter methods for caches in MarxanData and MarxanResults object.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = RaspData,name = character,y = ANY}: 

\item \code{x = RaspData,name = character,y = missing}: 

\item \code{x = RaspResults,name = character,y = ANY}: 

\item \code{x = RaspResults,name = character,y = missing}: 
}}
\note{
caches are implemented using environments, the hash is used as the name of the object in the environment.
}
\keyword{internal}

