% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raptr-internal.R, R/RapData.R, R/RapResults.R
\name{cache}
\alias{cache}
\title{Get and set cache methods}
\usage{
cache(x, name, y)

\S4method{cache}{RapData,character,ANY}(x, name, y)

\S4method{cache}{RapData,character,missing}(x, name, y)

\S4method{cache}{RapResults,character,ANY}(x, name, y)

\S4method{cache}{RapResults,character,missing}(x, name, y)
}
\arguments{
\item{x}{\code{RapData} or \code{RapResults} object}

\item{name}{\code{character} hash.}

\item{y}{if \code{ANY} this object gets cached with name, else if
\code{missing} the object hashed at name gets returned.}
}
\value{
\code{ANY} or \code{NULL} depends on \code{y} argument.
}
\description{
Getter and setter methods for caches in RapData and RapResults object.
}
\note{
caches are implemented using environments, the hash is used as the
name of the object in the environment.
}
\keyword{internal}
