% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{casestudy_data}
\alias{casestudy_data}
\alias{cs_pus}
\alias{cs_rs_amount}
\alias{cs_rs_space}
\alias{cs_ru}
\alias{cs_space}
\alias{cs_spp}
\title{Case-study dataset for a conservation planning exercise}
\format{\code{SpatialPolyognsDataFrame}, \code{RasterLayer}, \code{RasterLayer}.}
\usage{
cs_pus

cs_space

cs_spp

cs_ru

cs_rs_amount

cs_rs_space
}
\description{
This dataset contains data to generate example prioritisations for the pale-headed Rosella (\emph{Platycercus adscitus}) in Queensland, Australia. Specific objects in the dataset include:
\itemize{
	\item cs_pus: \code{SpatialPolygonsDataFrame} planning units. The units were generated as \eqn{10km^2} squares across the species' range, and then clipped to the Queensland, Australia (using data obtained from the \href{Australia Bureau of Statistics}{http://www.abs.gov.au/ausstats/abs@.nsf/mf/1259.0.30.001?OpenDocument}). They were then overliad with Australia's protected area network (obtained from the World Database on Protected Areas (WDPA) at \url{http://www.protectedplanet.net/}). This attribute table has 3 fields. The 'area' field contains the units' area, the 'cost' field is set to 1 for all units, and the 'status' filed indicates if 50\% or more of the units' extent is covered by protected areas.
	\item cs_spp: \code{RasterLayer} probability distribution map for the \emph{P. adscitus} clipped to Queensland, Australia. This map was dervived from records in the BirdLife Altas Database (\url{http://birdata.com.au/}).
	\item cs_space: \code{RasterStack} describing broad-scale variation across Queensland. containing the projections of the first two axes from a annual precipitation data across the species range (obtained from \url{http://www.worldclim.org/} as BIO12, and resampled to \eqn{10km^2} resolution).
	\item cs_ru: \code{RaspUnsolved} object.
	\item cs_rs_amount: \code{RaspSolved} object containing 5 solutions identified using amount-based targets.
	\item cs_rs_space: \code{RaspSolved} object containing 5 solutions identified using both amount-based and space-based targets.
}
}
\examples{
# load data
data(cs_pus, cs_spp, cs_space, cs_ru, cs_rs_amount, cs_rs_space)
# plot data
\dontrun{
plot(cs_pus)
plot(cs_spp)
plot(cs_space)
plot(cs_rs_amount)
plot(cs_rs_space)
}
}
\keyword{datasets}

