% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandPoints.R
\docType{class}
\name{DemandPoints-class}
\alias{DemandPoints-class}
\title{DemandPoints: An S4 class to represent demand points}
\description{
This class is used to store demand point information.
}
\section{Slots}{

\describe{
\item{\code{coords}}{\code{\link[base]{matrix}} of coordinates for each demand point.}

\item{\code{weights}}{\code{numeric} weights for each demand point.}
}}

\seealso{
\code{\link{DemandPoints}}
}
