% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DemandPoints.R
\name{DemandPoints}
\alias{DemandPoints}
\title{Create new DemandPoints object}
\usage{
DemandPoints(points, weights)
}
\arguments{
\item{points}{\code{SimplePoints} coordinates for each demand point in the attribute space.}

\item{weights}{\code{numeric} weights for each demand poont.}
}
\description{
This funciton creates a new DemandPoints object
}
\examples{
dps <- DemandPoints(
	SimplePoints(matrix(rnorm(100), ncol=2)),
	runif(50)
)
}
\seealso{
\code{\link{DemandPoints-class}}
}

