% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemandPoints.R
\name{DemandPoints}
\alias{DemandPoints}
\title{Create new DemandPoints object}
\usage{
DemandPoints(coords, weights)
}
\arguments{
\item{coords}{\code{matrix} of coordinates for each demand point.}

\item{weights}{\code{numeric} weights for each demand point.}
}
\description{
This function creates a new DemandPoints object
}
\examples{
dps <- DemandPoints(
matrix(rnorm(100), ncol=2),
runif(50)
)
}
\seealso{
\code{\link{DemandPoints-class}}
}

