% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GurobiOpts.R
\docType{class}
\name{GurobiOpts-class}
\alias{GurobiOpts-class}
\title{GurobiOpts: An S4 class to represent Gurobi parameters}
\description{
This class is used to store Gurobi input parameters.
}
\section{Slots}{

\describe{
\item{\code{Threads}}{\code{integer} number of cores to use for processing. Defaults to 1L.}

\item{\code{MIPGap}}{\code{numeric} MIP gap specifying minimum solution quality. Defaults to 0.05.}

\item{\code{Presolve}}{\code{integer} code for level of computation in presolve. Defaults to 2.}

\item{\code{TimeLimit}}{\code{integer} number of seconds to allow for solving. Defaults to NA_integer_, and so a time limit is not imposed.}

\item{\code{NumberSolutions}}{\code{integer} number of solutions to generate. Defaults to 1L.}
}}

