% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GurobiOpts.R
\docType{class}
\name{GurobiOpts-class}
\alias{GurobiOpts-class}
\title{GurobiOpts: An S4 class to represent Gurobi parameters}
\description{
This class is used to store Gurobi input parameters.
}
\section{Slots}{

\describe{
\item{\code{Threads}}{\code{integer} number of cores to use for processing. Defaults to 1L.}

\item{\code{MIPGap}}{\code{numeric} MIP gap specifying minimum solution quality. Defaults to 0.1.}

\item{\code{Method}}{\code{integer} Algorithm to use for solving model. Defaults to 0L.}

\item{\code{Presolve}}{\code{integer} code for level of computation in presolve. Defaults to 2.}

\item{\code{TimeLimit}}{\code{integer} number of seconds to allow for solving. Defaults to NA_integer_, and so a time limit is not imposed.}

\item{\code{NumberSolutions}}{\code{integer} number of solutions to generate. Defaults to 1L.}

\item{\code{MultipleSolutionsMethod}}{\code{character} name of method to obtain multiple solutions \code{NumberSolutions} > 1. Available options are 'benders.cuts' and 'solution.pool'. Defaults to 'benders.cuts'. Note that the \code{rgurobi} package must be to use the 'solution.pool' method.}
}}
\seealso{
\code{\link{GurobiOpts}}.
}

