% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GurobiOpts.R
\name{GurobiOpts}
\alias{GurobiOpts}
\title{Create GurobiOpts object}
\usage{
GurobiOpts(Threads = 1L, MIPGap = 0.05, Presolve = 2L,
  TimeLimit = NA_integer_)
}
\arguments{
\item{Threads}{\code{integer} number of cores to use for processing. Defaults to 1L.}

\item{MIPGap}{\code{numeric} MIP gap (lp_solve parameter). Defaults to 0.05.}

\item{Presolve}{\code{integer} code for level of computation in presolve (lp_solve parameter). Defaults to 2.}

\item{TimeLimit}{\code{integer} number of seconds to allow for solving. Defaults to NA_integer_, and so a time limit is not imposed.}
}
\value{
\code{GurobiOpts} object
}
\description{
This function creates a new GurobiOpts object.
}
\examples{
# create GurobiOpts object using default parameters
x <- GurobiOpts(Threads=1L, MIPGap=0.05, Presolve=2L, TimeLimit=NA_integer_)
# print object
print(x)
}
\seealso{
\code{\link{GurobiOpts-class}}.
}

