% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raptr-internal.R, R/RapData.R, R/RapResults.R
\name{is.cached}
\alias{is.cached}
\title{Test if hash is cached in a Rap object}
\arguments{
\item{x}{\code{RapData} or \code{RapResults} object}

\item{name}{\code{character} hash.}
}
\value{
\code{logical} Is it cached?
}
\description{
Tests if hash is cached in Rap object.
}
\note{
caches are implemented using environments, the hash is used as the
  name of the object in the environment.
}
\keyword{internal}
