% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R, R/RaspSolved.R, R/generics.R
\docType{methods}
\name{is.comparable,RaspData,RaspData-method}
\alias{is.comparable}
\alias{is.comparable,RaspData,RaspData-method}
\alias{is.comparable,RaspData,RaspUnsolvedOrSolved-method}
\alias{is.comparable,RaspUnsolvedOrSolved,RaspData-method}
\alias{is.comparable,RaspUnsolvedOrSolved,RaspUnsolvedOrSolved-method}
\title{Compare Rasp objects}
\usage{
\S4method{is.comparable}{RaspData,RaspData}(x, y)

\S4method{is.comparable}{RaspUnsolvedOrSolved,RaspUnsolvedOrSolved}(x, y)

\S4method{is.comparable}{RaspData,RaspUnsolvedOrSolved}(x, y)

\S4method{is.comparable}{RaspUnsolvedOrSolved,RaspData}(x, y)

is.comparable(x, y)
}
\arguments{
\item{x}{\code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.}

\item{y}{\code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.}
}
\value{
\code{logical} are the objects based on the same data?
}
\description{
This function checks objects to see if they share the same input data.
}
\seealso{
\code{\link{RaspData-class}}, \code{\link{RaspUnsolved-class}}, \code{\link{RaspSolved-class}}.
}

