% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raptr-internal.R, R/RapData.R, R/RapSolved.R
\docType{methods}
\name{is.comparable}
\alias{is.comparable}
\alias{is.comparable,RapData,RapData-method}
\alias{is.comparable,RapUnsolOrSol,RapUnsolOrSol-method}
\alias{is.comparable,RapData,RapUnsolOrSol-method}
\alias{is.comparable,RapUnsolOrSol,RapData-method}
\title{Compare Rap objects}
\usage{
\S4method{is.comparable}{RapData,RapData}(x, y)

\S4method{is.comparable}{RapUnsolOrSol,RapUnsolOrSol}(x, y)

\S4method{is.comparable}{RapData,RapUnsolOrSol}(x, y)

\S4method{is.comparable}{RapUnsolOrSol,RapData}(x, y)
}
\arguments{
\item{x}{\code{RapData}, \code{RapUnsolved}, or \code{RapSolved} object.}

\item{y}{\code{RapData}, \code{RapUnsolved}, or \code{RapSolved} object.}
}
\value{
\code{logical} are the objects based on the same data?
}
\description{
This function checks objects to see if they share the same input data.
}
\seealso{
\code{\link{RapData-class}}, \code{\link{RapUnsolved-class}},
\code{\link{RapSolved-class}}.
}
\keyword{internal}
