% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.GurobiInstalled}
\alias{is.GurobiInstalled}
\title{Test if Gurobi is installed on computer}
\usage{
is.GurobiInstalled(verbose = TRUE)
}
\arguments{
\item{verbose}{\code{logical} should messages be printed?}
}
\value{
\code{logical} Is it installed and ready to use?
}
\description{
This function determines if the Gurobi R package is installed on the
computer and that it can be used \code{\link[base]{options}}.
}
\examples{
\donttest{
# check if Gurobi is installed
is.GurobiInstalled()

# print cached status of installation
options()$GurobiInstalled
}

}
\seealso{
\code{\link[base]{options}}.
}
