% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R
\name{make.RaspData}
\alias{make.RaspData}
\title{Make data for RASP using minimal inputs}
\usage{
make.RaspData(pus, species, spaces = NULL, area.targets = 0.2,
  space.targets = 0.2, n.demand.points = 1000L,
  kernel.method = c("sm.density", "hyperbox")[1], quantile = 0.2,
  species.points = NULL, n.species.points = ceiling(0.2 * cellStats(species,
  "sum")), include.geographic.space = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{pus}{\code{SpatialPolygons} with planning unit data.}

\item{species}{\code{RasterLayer}, \code{RasterStack}, \code{RasterBrick} with species probability distribution data.}

\item{spaces}{\code{list} of/or \code{RasterLayer}, \code{RasterStack}, \code{RasterBrick} representing projects of attribute space over geographic space. Use a \code{list} to denote seperate attribute spaces.}

\item{area.targets}{\code{numeric} vector for area targets (\%) for each species. Defaults to 0.2 for each attribute space for each species.}

\item{space.targets}{\code{numeric} vector for attribute space targets (\%) for each species. Defaults to 0.2 for each attribute space for each species. Note all attribute spaces have the same targets.}

\item{n.demand.points}{\code{integer} number of demand points to use for each attribute space for each species.}

\item{kernel.method}{\code{character} name of kernel method to use to generate demand points. Use either \code{sm.density} or \code{hypervolume}.}

\item{quantile}{\code{numeric} quantile to generate demand points within. If 0 then demand points are generated across the full range of values the \code{species.points} intersect. Defaults to 0.2.}

\item{species.points}{\code{list} of/or \code{SpatialPointsDataFrame} or \code{SpatialPoints} with species presence records. Use a \code{list} of objects to represent different species. Must have the same number of elements as \code{species}. If not supplied then use \code{n.species.points} to sample points from the species distributions.}

\item{n.species.points}{\code{numeric} vector specfiying the number points to sample the species distributions to use to generate demand points. Defaults to 20\% of the distribution.}

\item{include.geographic.space}{\code{logical} should the geographic space be considered an attribute space?}

\item{verbose}{\code{logical} print statements during processing?}

\item{...}{additional arguments to \code{calcBoundaryData} and \code{calcPuVsSpeciesData}.}
}
\description{
This function prepares spatially explicit planning unit, species data, and landscape data layers for RASP processing.
}
\examples{
data(pus, species, space)
x<-RaspData(pus, species, space)
}
\seealso{
\code{\link{RaspData-class}}, \code{\link{RaspData}}.
}

