% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManualOpts.R
\name{ManualOpts}
\alias{ManualOpts}
\title{Create ManualOpts object}
\usage{
ManualOpts(NumberSolutions = 1L)
}
\arguments{
\item{NumberSolutions}{\code{integer} number of solutions to generate.
Defaults to 1L.}
}
\value{
\code{\link{ManualOpts}} object
}
\description{
This function creates a new ManualOpts object.
}
\examples{
# create ManualOpts object
ManualOpts(NumberSolutions = 1L)

}
\seealso{
\code{\link{ManualOpts-class}}.
}
