% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspSolved.R, R/generics.R
\docType{methods}
\name{plot,RaspSolved,numeric-method}
\alias{plot}
\alias{plot,RaspSolved,RaspSolved-method}
\alias{plot,RaspSolved,missing-method}
\alias{plot,RaspSolved,numeric-method}
\title{Plot RASP object}
\usage{
\S4method{plot}{RaspSolved,numeric}(x, y, basemap = "none",
  color.palette = "Greens", locked.in.color = "#000000FF",
  locked.out.color = "#D7D7D7FF", alpha = ifelse(basemap == "none", 1, 0.7),
  grayscale = FALSE, force.reset = FALSE)

\S4method{plot}{RaspSolved,missing}(x, y, basemap = "none",
  color.palette = "PuBu", locked.in.color = "#000000FF",
  locked.out.color = "#D7D7D7FF", alpha = ifelse(basemap == "none", 1, 0.7),
  grayscale = FALSE, force.reset = FALSE)

\S4method{plot}{RaspSolved,RaspSolved}(x, y, i = NULL, j = i,
  basemap = "none", color.palette = ifelse(is.null(i), "RdYlBu", "Accent"),
  x.locked.in.color = "#000000FF", x.locked.out.color = "#D7D7D7FF",
  y.locked.in.color = "#FFFFFFFF", y.locked.out.color = "#D7D7D7FF",
  alpha = ifelse(basemap == "none", 1, 0.7), grayscale = FALSE,
  force.reset = FALSE)
}
\arguments{
\item{x}{\code{RaspSolved} object.}

\item{y}{\code{NULL} to plot selection frequencies. \code{numeric} to plot the i'th solution, or 0 to plot the best solution. \code{RaspSolved} object to plot differences in solutions between objects. Defaults to \code{ULL}.}

\item{basemap}{\code{character} object indicating the type of basemap to use (see \code{link{basemap}}). Use either 'none', 'roadmap', 'mobile', 'satellite', 'terrain', 'hybrid', 'mapmaker-roadmap', 'mapmaker-hybrid'. Defaults to 'none'.}

\item{color.palette}{\code{character} name of colour palette to use for planning units (see \code{\link[RColorBrewer]{brewer.pal}}). Defaults to 'PuBu' when \code{y} is \code{NULL}, 'Greens' when \code{y} is \code{numeric}, and 'RdYlBu' or 'Accent' when \code{y} is \code{RaspSolved}.}

\item{locked.in.color}{\code{character} color to denote locked in planning units. Used when \code{y} is \code{NULL}. Defaults to '#000000FF'.}

\item{locked.out.color}{\code{character} color to denote locked in planning units. Used when \code{y} is \code{NULL}. Defaults to '#D7D7D7FF'.}

\item{alpha}{\code{numeric} value to indicate how transparent the planning unit colors shoud be.}

\item{grayscale}{\code{logical} should the basemap be gray-scaled?}

\item{force.reset}{\code{logical} if basemap data has been cached, should it be re-downloaded?}

\item{i}{\code{NULL} to plot selection frequencies. \code{numeric} to plot the i'th solution, or 0 to plot the best solution. Only used when \code{y} is a \code{RaspSolved} object. Defaults to \code{NULL}.}

\item{j}{\code{NULL} to plot selection frequencies. \code{numeric} to plot the i'th solution, or 0 to plot the best solution. Only used when \code{y} is a \code{RaspSolved} object. Defaults to \code{j}.}

\item{x.locked.in.color}{\code{character} color to denote locked in planning units in \code{x}. Used when \code{y} is \code{RaspSolved}. Defaults to '#000000FF'.}

\item{x.locked.out.color}{\code{character} color to denote locked out planning units in \code{x}. Used when \code{y} is \code{RaspSolved}. Defaults to '#D7D7D7FF'.}

\item{y.locked.in.color}{\code{character} color to denote locked in planning units in \code{y}. Used when \code{y} is \code{RaspSolved}. Defaults to '#FFFFFFFF'.}

\item{y.locked.out.color}{\code{character} color to denote locked out planning units in \code{y}. Used when \code{y} is \code{RaspSolved}. Defaults to '#D7D7D7FF'.}
}
\description{
This function plots the solutions contained in \code{RaspSolved} objects. It can be used to show a single solution, or the the selection frequencies of planning
units contained in a single \code{RaspSolved} object. Additionally, two \code{RaspSolved} objects can be supplied to plot the differences between them.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = RaspSolved,y = numeric}: 

\item \code{x = RaspSolved,y = missing}: 

\item \code{x = RaspSolved,y = RaspSolved}: 
}}

