% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/RapSolved.R
\name{plot}
\alias{plot}
\alias{plot,RapSolved,numeric-method}
\alias{plot,RapSolved,missing-method}
\alias{plot,RapSolved,RapSolved-method}
\title{Plot object}
\usage{
\S4method{plot}{RapSolved,numeric}(x, y, basemap = "none",
 pu.color.palette = c("#e5f5f9", "#00441b", "#FFFF00", "#FF0000"), alpha =
 ifelse(basemap == "none", 1, 0.7), grayscale = FALSE, main = NULL,
 force.reset = FALSE)

\S4method{plot}{RapSolved,missing}(x, y, basemap = "none",
pu.color.palette = c("PuBu", "#FFFF00", "#FF0000"),
alpha = ifelse(basemap == "none", 1, 0.7),
grayscale = FALSE, main = NULL,
force.reset = FALSE)

\S4method{plot}{RapSolved,RapSolved}(x, y, i = NULL, j = i,
basemap = "none",
pu.color.palette = ifelse(is.null(i), c("RdYlBu", "#FFFF00",
"#FF0000"), "Accent"),
alpha = ifelse(basemap == "none", 1, 0.7),
grayscale = FALSE, main = NULL, force.reset = FALSE)
}
\arguments{
\item{x}{\code{\link[=RapSolved]{RapSolved()}} object.}

\item{y}{Available inputs are: \code{NULL} to plot selection frequencies,
\code{numeric} number to plot a specific solution,
\code{0} to plot the best solution, and a \code{\link[=RapSolved]{RapSolved()}}
object to plot differences in solutions between objects. Defaults to
\code{NULL}.}

\item{basemap}{\code{character} object indicating the type of basemap to use
(see \code{\link[=basemap]{basemap()}}). Valid options include \code{"none"},
\code{"roadmap"}, \code{"mobile"}, \code{"satellite"}, \code{"terrain"},
\code{"hybrid"}, \code{"mapmaker-roadmap"}, \code{"mapmaker-hybrid"}.
Defaults to \code{"none"} such that no basemap is shown.}

\item{pu.color.palette}{\code{character} vector of colors to indicate
planning unit statuses.
If plotting selection frequencies (i.e., \code{j = NULL}), then
defaults to a \code{c("PuBu", "#FFFF00", "#FF0000")}.
Here, the first element corresponds to a color palette
(per \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}}) and the last two elements
indicate the colors for locked in and locked out planning units.
Otherwise, the parameter defaults to a \code{character} vector of
\code{c("grey30", "green", "yellow", "black", "gray80", "red", "orange")}.}

\item{alpha}{\code{numeric} value to indicating the transparency level for
coloring the planning units.}

\item{grayscale}{\code{logical} should the basemap be gray-scaled?}

\item{main}{\code{character} title for the plot. Defaults to \code{NULL} and
a default title is used.}

\item{force.reset}{\code{logical} if basemap data has been cached, should it
be re-downloaded?}

\item{i}{Available inputs are: \code{NULL} to plot selection frequencies.
\code{numeric} to plot a specific solution, \code{0} to plot the best
solution. This argument is only used when \code{y} is a
\code{\link[=RapSolved]{RapSolved()}} object. Defaults to \code{NULL}.}

\item{j}{Available inputs are: \code{NULL} to plot selection frequencies.
\code{numeric} to plot a specific solution, \code{0} to plot the best
solution. This argument is only used when \code{y} is a
\code{\link[=RapSolved]{RapSolved()}} object. Defaults to argument \code{j}.}
}
\description{
This function plots the solutions contained in \code{\link[=RapSolved]{RapSolved()}}
objects. It can be used to show a single solution, or the the selection
frequencies of planning units contained in a single \code{\link[=RapSolved]{RapSolved()}}
object. Additionally, two \code{\link[=RapSolved]{RapSolved()}} objects can be supplied to
plot the differences between them.
}
\details{
This function requires the \pkg{RgoogleMaps} package to be installed
in order to create display a basemap.
}
\examples{
# load example data set with solutions
data(sim_rs)

# plot selection frequencies
plot(sim_rs)

# plot best solution
plot(sim_rs, 0)

# plot second solution
plot(sim_rs, 2)

# plot different between best and second solutions
plot(sim_rs, sim_rs, 0 ,2)
}
\seealso{
\code{\link[=RapSolved]{RapSolved()}}.
}
