% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GurobiOpts.R, R/RaspData.R, R/RaspOpts.R, R/RaspResults.R, R/RaspSolved.R, R/RaspUnsolved.R, R/generics.R
\name{print.GurobiOpts}
\alias{print}
\alias{print.GurobiOpts}
\alias{print.RaspData}
\alias{print.RaspOpts}
\alias{print.RaspResults}
\alias{print.RaspSolved}
\alias{print.RaspUnsolved}
\title{Print objects}
\usage{
\method{print}{GurobiOpts}(x, ..., header = TRUE)

\method{print}{RaspData}(x, ..., header = TRUE)

\method{print}{RaspOpts}(x, ..., header = TRUE)

\method{print}{RaspResults}(x, ..., header = TRUE)

\method{print}{RaspSolved}(x, ...)

\method{print}{RaspUnsolved}(x, ...)
}
\arguments{
\item{x}{\code{GurobiOpts}, \code{RaspOpts}, \code{RaspData}, \code{RaspUnsolved}, \code{RaspResults}, or \code{RaspSolved} object.}

\item{...}{not used.}

\item{header}{\code{logical} should object header be included?}
}
\description{
Prints objects.
}

