% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GurobiOpts.R, R/ManualOpts.R, R/RaspData.R, R/RaspReliableOpts.R, R/RaspResults.R, R/RaspSolved.R, R/RaspUnreliableOpts.R, R/RaspUnsolved.R, R/generics.R
\name{print.GurobiOpts}
\alias{print}
\alias{print.GurobiOpts}
\alias{print.ManualOpts}
\alias{print.RaspData}
\alias{print.RaspReliableOpts}
\alias{print.RaspResults}
\alias{print.RaspSolved}
\alias{print.RaspUnreliableOpts}
\alias{print.RaspUnsolved}
\title{Print objects}
\usage{
\method{print}{GurobiOpts}(x, ..., header = TRUE)

\method{print}{ManualOpts}(x, ..., header = TRUE)

\method{print}{RaspData}(x, ..., header = TRUE)

\method{print}{RaspReliableOpts}(x, ..., header = TRUE)

\method{print}{RaspResults}(x, ..., header = TRUE)

\method{print}{RaspSolved}(x, ...)

\method{print}{RaspUnreliableOpts}(x, ..., header = TRUE)

\method{print}{RaspUnsolved}(x, ...)
}
\arguments{
\item{x}{\code{GurobiOpts}, \code{RaspUnreliableOpts}, \code{RaspReliableOpts}, \code{RaspData}, \code{RaspUnsolved}, \code{RaspResults}, or \code{RaspSolved} object.}

\item{...}{not used.}

\item{header}{\code{logical} should object header be included?}
}
\description{
Prints objects.
}
\examples{
# load data
data(sim_ru, sim_rs)
## print classes in package
# GurobiOpts
print(GurobiOpts())
# RaspReliableOpts
print(RaspReliableOpts())
# RaspUnreliableOpts
print(RaspUnreliableOpts())
# RaspData
print(sim_ru@data)
# RaspUnsolved
print(sim_ru)
# RaspResults
print(sim_rs@results)
# RaspSolved
print(sim_rs)
}
\seealso{
\code{\link{GurobiOpts}}, \code{\link{RaspUnreliableOpts}}, \code{\link{RaspReliableOpts}}, \code{\link{RaspData}}, \code{\link{RaspUnsolved}}, \code{\link{RaspResults}}, \code{\link{RaspSolved}}.
}

