% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/AttributeSpace.R, R/AttributeSpaces.R, R/GurobiOpts.R, R/ManualOpts.R, R/RapData.R, R/RapReliableOpts.R, R/RapResults.R, R/RapUnreliableOpts.R, R/RapUnsolved.R, R/RapSolved.R
\name{print}
\alias{print}
\alias{print.AttributeSpace}
\alias{print.AttributeSpaces}
\alias{print.GurobiOpts}
\alias{print.ManualOpts}
\alias{print.RapData}
\alias{print.RapReliableOpts}
\alias{print.RapResults}
\alias{print.RapSolved}
\alias{print.RapUnreliableOpts}
\alias{print.RapUnsolved}
\title{Print objects}
\usage{
\method{print}{AttributeSpace}(x, ..., header = TRUE)

\method{print}{AttributeSpaces}(x, ..., header = TRUE)

\method{print}{GurobiOpts}(x, ..., header = TRUE)

\method{print}{ManualOpts}(x, ..., header = TRUE)

\method{print}{RapData}(x, ..., header = TRUE)

\method{print}{RapReliableOpts}(x, ..., header = TRUE)

\method{print}{RapResults}(x, ..., header = TRUE)

\method{print}{RapUnreliableOpts}(x, ..., header = TRUE)

\method{print}{RapUnsolved}(x, ...)

\method{print}{RapSolved}(x, ...)
}
\arguments{
\item{x}{\code{GurobiOpts}, \code{RapUnreliableOpts}, \code{RapReliableOpts}, \code{RapData}, \code{RapUnsolved}, \code{RapResults}, or \code{RapSolved} object.}

\item{...}{not used.}

\item{header}{\code{logical} should object header be included?}
}
\description{
Prints objects.
}
\examples{
# load data
data(sim_ru, sim_rs)
## print classes in package
# GurobiOpts
print(GurobiOpts())
# RapReliableOpts
print(RapReliableOpts())
# RapUnreliableOpts
print(RapUnreliableOpts())
# RapData
print(sim_ru@data)
# RapUnsolved
print(sim_ru)
# RapResults
print(sim_rs@results)
# RapSolved
print(sim_rs)
}
\seealso{
\code{\link{GurobiOpts}}, \code{\link{RapUnreliableOpts}}, \code{\link{RapReliableOpts}}, \code{\link{RapData}}, \code{\link{RapUnsolved}}, \code{\link{RapResults}}, \code{\link{RapSolved}}.
}

