% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R, R/RaspSolved.R, R/RaspUnsolved.R, R/generics.R
\name{pu.subset.RaspData}
\alias{pu.subset}
\alias{pu.subset.RaspData}
\alias{pu.subset.RaspSolved}
\alias{pu.subset.RaspUnsolved}
\title{Subset planning units}
\usage{
\method{pu.subset}{RaspData}(x, pu)

\method{pu.subset}{RaspSolved}(x, pu)

\method{pu.subset}{RaspUnsolved}(x, pu)

pu.subset(x, pu)
}
\arguments{
\item{x}{\code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.}

\item{pu}{\code{integer} vector to specify the index of planning units to subset.}
}
\value{
\code{RaspData} or \code{RaspUnsolved} object depending on input object.
}
\description{
Subset planning units from a \code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.
}
\seealso{
\code{RaspData}, \code{RaspUnsolved}, \code{RaspSolved}.
}

