% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/RapData.R, R/RapSolved.R
\name{pu.subset}
\alias{pu.subset}
\alias{pu.subset.RapData}
\alias{pu.subset.RapUnsolOrSol}
\title{Subset planning units}
\usage{
pu.subset(x, pu)

\method{pu.subset}{RapData}(x, pu)

\method{pu.subset}{RapUnsolOrSol}(x, pu)
}
\arguments{
\item{x}{\code{\link[=RapData]{RapData()}}, \code{\link[=RapUnsolved]{RapUnsolved()}}, or
\code{\link[=RapSolved]{RapSolved()}} object.}

\item{pu}{\code{integer} vector to specify the index of planning units to
subset.}
}
\value{
\code{\link[=RapData]{RapData()}} or \code{\link[=RapUnsolved]{RapUnsolved()}} object depending
on input object.
}
\description{
Subset planning units from a \code{\link[=RapData]{RapData()}},
\code{\link[=RapUnsolved]{RapUnsolved()}}, or \code{\link[=RapSolved]{RapSolved()}} object.
}
\examples{
\dontrun{
# load data
data(sim_ru)

# generate new object with first 10 planning units
sim_ru2 <- pu.subset(sim_ru, seq_len(10))
}
}
\seealso{
\code{\link[=RapData]{RapData()}}, \code{\link[=RapUnsolved]{RapUnsolved()}},
\code{\link[=RapSolved]{RapSolved()}}.
}
