% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RapReliableOpts.R
\name{RapReliableOpts}
\alias{RapReliableOpts}
\title{Create RapReliableOpts object}
\usage{
RapReliableOpts(BLM = 0, FAILUREMULTIPLIER = 1.1, MAXRLEVEL = 5L)
}
\arguments{
\item{BLM}{\code{numeric} boundary length modifier. Defaults to 0.}

\item{FAILUREMULTIPLIER}{\code{numeric} multiplier for failure planning unit. Defaults to 1.1.}

\item{MAXRLEVEL}{\code{numeric} maximum R failure level for approximation. Defaults to 5L.}
}
\value{
\code{RapReliableOpts} object
}
\description{
This function creates a new RapReliableOpts object.
}
\examples{
# create RapReliableOpts using defaults
RapReliableOpts(BLM=0, FAILUREMULTIPLIER=1.1, MAXRLEVEL=5L)
}
\seealso{
\code{\link{RapReliableOpts-class}}.
}

