% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapResults.R
\docType{class}
\name{RapResults-class}
\alias{RapResults-class}
\title{RapResults: An S4 class to represent RAP results}
\description{
This class is used to store RAP results.
}
\details{
The \code{summary} table follows Marxan conventions (summary.dat in \url{http://marxan.net/downloads/uq_marxan_web_2/module5.html}). The columns are:
\tabular{cl}{
\code{Run_Number}
\tab The index of each solution in the object. \cr
\code{Status}
\tab The status of the solution. The values in this column correspond to outputs from the Gurobi software package (\url{http://www.gurobi.com/documentation/6.5/refman/optimization_status_codes.html}).\cr
\code{Score}
\tab The objective function for the solution. \cr
\code{Cost}
\tab Total cost associated with a solution \cr
\code{Planning_Units}
\tab Number of planning units selected in a solution \cr
\code{Connectivity_Total}
\tab The total amount of shared boundary length between all planning units. All solutions in the same object should have equal values for this column. \cr
\code{Connectivity_In}
\tab The amount of shared boundary length among planning units selected in the solution \cr
\code{Connectivity_Edge}
\tab The amount of exposed boundary length in the solution \cr
\code{Connectivity_Out}
\tab The number of shared boundary length among planning units not selected in the solution  \cr
\code{Connectivity_Fraction}
\tab The ratio of shared boundary length in the solution (\code{Connectivity_In}) to the total amount of boundary length (\code{Connectivity_Edge}). This ratio is an indicator of solution quality. Solutions with a lower ratio will have less planning units and will be more efficient. \cr
}
}
\section{Slots}{

\describe{
\item{\code{summary}}{\code{data.frame} with summary information on solutions.}

\item{\code{selections}}{\code{matrix} with binary selections. The cell \eqn{x_{ij}} denotes if planning unit \eqn{j} is selected in the \eqn{i}'th solution}

\item{\code{amount.held}}{\code{matrix} with the amount held for each species in each solution.}

\item{\code{space.held}}{\code{matrix} with the proportion of attribute space sampled for each species in each solution.}

\item{\code{best}}{\code{integer} with index of best solution.}

\item{\code{logging.file}}{\code{character} Gurobi log files.}

\item{\code{.cache}}{\code{environment} used to store extra data.}
}}
\seealso{
\code{\link{RapResults}}, \code{\link{read.RapResults}}.
}

