% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapSolved.R
\docType{class}
\name{RapSolved-class}
\alias{RapSolved-class}
\title{RapSolved: An S4 class to represent RAP inputs and outputs}
\description{
This class is used to store RAP input and output data in addition to input
parameters.
}
\section{Slots}{

\describe{
\item{\code{opts}}{\code{\link[=RapReliableOpts]{RapReliableOpts()}} or \code{\link[=RapUnreliableOpts]{RapUnreliableOpts()}}
object used to store input parameters.}

\item{\code{solver}}{\code{\link[=GurobiOpts]{GurobiOpts()}} or \code{\link[=ManualOpts]{ManualOpts()}} object
used to store solver information/parameters.}

\item{\code{data}}{\code{\link[=RapData]{RapData()}} object used to store input data.}

\item{\code{results}}{\code{\link[=RapResults]{RapResults()}} object used to store results.}
}}

\seealso{
\linkS4class{RapReliableOpts},
\linkS4class{RapUnreliableOpts}, \linkS4class{RapData},
\linkS4class{RapResults}.
}
