% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapSolved.R
\name{RapSolved-class}
\alias{RapSolved-class}
\title{RapSolved: An S4 class to represent RAP inputs and outputs}
\description{
This class is used to store RAP input and output data in addition to input
parameters.
}
\section{Slots}{

\describe{
\item{\code{opts}}{\code{\link{RapReliableOpts}} or \code{\link{RapUnreliableOpts}}
object used to store input parameters.}

\item{\code{solver}}{\code{\link{GurobiOpts}} or \code{\link{ManualOpts}} object
used to store solver information/parameters.}

\item{\code{data}}{\code{\link{RapData}} object used to store input data.}

\item{\code{results}}{\code{\link{RapResults}} object used to store results.}
}}

\seealso{
\code{\link{RapReliableOpts-class}},
  \code{\link{RapUnreliableOpts-class}}, \code{\link{RapData-class}},
  \code{\link{RapResults-class}}.
}
