% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapSolved.R
\name{RapSolved}
\alias{RapSolved}
\title{Create new RapSolved object}
\usage{
RapSolved(unsolved, solver, results)
}
\arguments{
\item{unsolved}{\code{\link[=RapUnsolved]{RapUnsolved()}} object.}

\item{solver}{\code{\link[=GurobiOpts]{GurobiOpts()}} or \code{\link[=ManualOpts]{ManualOpts()}} object.}

\item{results}{\code{\link[=RapResults]{RapResults()}} object.}
}
\value{
\code{\link[=RapSolved]{RapSolved()}} object.
}
\description{
This function creates a \code{\link[=RapSolved]{RapSolved()}} object.
}
\seealso{
\linkS4class{RapSolved}, \linkS4class{RapResults},
\code{link{solve}}.
}
