% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapSolved.R
\name{RapSolved}
\alias{RapSolved}
\title{Create new RapSolved object}
\usage{
RapSolved(unsolved, solver, results)
}
\arguments{
\item{unsolved}{\code{RapUnsolved} object.}

\item{solver}{\code{GurobiOpts} or \code{ManualOpts} object.}

\item{results}{\code{RapResults} object.}
}
\value{
\code{RapSolved} object.
}
\description{
This function creates a \code{RapSolved} object.
}
\seealso{
\code{\link{RapSolved-class}}, \code{\link{RapResults-class}}, \code{link{solve}}.
}

