% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapUnsolved.R
\name{RapUnsolved-class}
\alias{RapUnsolved-class}
\title{RapUnsolved: An S4 class to represent RAP inputs}
\description{
This class is used to store RAP input data and input parameters.
}
\section{Slots}{

\describe{
\item{\code{opts}}{\code{\link{RapReliableOpts}} or \code{\link{RapUnreliableOpts}}
object used to store input parameters.}

\item{\code{data}}{\code{\link{RapData}} object used to store input data.}
}}

\seealso{
\code{\link{RapReliableOpts-class}},
  \code{\link{RapUnreliableOpts-class}}, \code{\link{RapData-class}}.
}
