% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapUnsolved.R
\name{RapUnsolved}
\alias{RapUnsolved}
\title{Create a new RapUnsolved object}
\usage{
RapUnsolved(opts, data)
}
\arguments{
\item{opts}{\code{\link[=RapReliableOpts]{RapReliableOpts()}} or \code{\link[=RapUnreliableOpts]{RapUnreliableOpts()}}
object.}

\item{data}{\code{\link[=RapData]{RapData()}} object.}
}
\value{
\code{\link[=RapUnsolved]{RapUnsolved()}} object.
}
\description{
This function creates a \code{\link[=RapUnsolved]{RapUnsolved()}} object using a
\code{\link[=GurobiOpts]{GurobiOpts()}}, a \code{\link[=RapReliableOpts]{RapReliableOpts()}} or
\code{\link[=RapUnreliableOpts]{RapUnreliableOpts()}} object, and a \code{\link[=RapData]{RapData()}} object.
}
\examples{
\dontrun{
# set random number generator seed
set.seed(500)

# load data
cs_pus <- sf::read_sf(
 system.file("extdata", "cs_pus.gpkg", package = "raptr")
)
cs_spp <- terra::rast(
  system.file("extdata", "cs_spp.tif", package = "raptr")
)

# create inputs for RapUnsolved
ro <- RapUnreliableOpts()
rd <- make.RapData(cs_pus[seq_len(10), ], cs_spp, NULL,
                   include.geographic.space = TRUE,n.demand.points = 5L)

# create RapUnsolved object
ru <- RapUnsolved(ro, rd)

# print object
print(ru)
}
}
\seealso{
\linkS4class{RapReliableOpts},
\linkS4class{RapUnreliableOpts}, \linkS4class{RapData}.
}
