% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R
\name{RaspData}
\alias{RaspData}
\title{Create new RaspData object}
\usage{
RaspData(pu, pu.species.probabilities, pu.points, species, demand.points,
  boundary, polygons = NULL, skipchecks = FALSE, .cache = new.env(), ...)
}
\value{
RaspData object
}
\description{
This function creates a "RaspData" object using pre-processed data.
}
\section{Slots}{

\describe{
\item{\code{polygons}}{\code{PolySet} planning unit spatial data or \code{NULL} if data not available.}

\item{\code{pu}}{\code{data.frame} planning unit data. Columns are 'id' (\code{integer}), 'cost' (\code{numeric}), 'area' (\code{numeric}), and 'status' (\code{integer}).}

\item{\code{pu.points}}{\code{list} of \code{matrix} objects with coordinates of the planning units in the attribute spaces.}

\item{\code{species}}{\code{data.frame} with species data. Columns are 'id' (\code{integer}), 'area.target' (\code{numeric}), 'space.target' (\code{numeric}), and 'name' (\code{character}, optional).}

\item{\code{demand.points}}{\code{list} of \code{DemandPoint} objects with the demand points.}

\item{\code{pu.species.probabilities}}{\code{data.frame} with data on the probability of species in each planning unit. Columns are 'species' (\code{integer}), 'pu' (\code{integer}), and 'probability' (\code{numeric}) columns.}

\item{\code{boundary}}{\code{data.frame} with data on the shared boundary length of planning units. Columns are with 'id1' (\code{integer}), 'id2' (\code{integer}), and 'boundary' (\code{integer}).}

\item{\code{skipchecks}}{\code{logical} Skip data integrity checks? May improve speed for big data sets.}

\item{\code{.cache}}{\code{environment} used to cache calculations.}
}}
\note{
Generally, users are not encouraged to change arguments to \code{.cache}.
}
\seealso{
\code{\link[PBSmapping]{PolySet}}, \code{\link[sp]{SpatialPoints}}, \code{\link[sp]{SpatialPointsDataFrame}}, \code{\link{format.RaspData}}, \code{\link{RaspData-class}}.
}

