% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspOpts.R
\docType{class}
\name{RaspOpts-class}
\alias{RaspOpts-class}
\title{RaspOpts: An S4 class to represent RASP input parameters}
\description{
This class is used to store RASP input parameters.
}
\section{Slots}{

\describe{
\item{\code{BLM}}{\code{numeric} boundary length modifier. Defaults to 0.}

\item{\code{FAILUREMULTIPLIER}}{\code{numeric} multiplier for failure planning unit. Defaults to 1.1.}

\item{\code{MAXRLEVEL}}{\code{numeric} maximum R failure level for approximation. Defaults to 5L.}

\item{\code{NUMREPS}}{\code{integer} number of replicate runs. Defaults to 1L.}
}}

