% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspOpts.R
\name{RaspOpts}
\alias{RaspOpts}
\title{Create RaspOpts object}
\usage{
RaspOpts(BLM = 0, NUMREPS = 1L, NTHREADS = 1L, GAP = 0.05,
  PRESOLVE = 2L, TIMELIMIT = NA_integer_, VERBOSITY = 1L)
}
\arguments{
\item{BLM}{\code{numeric} boundary length modifier. Defaults to 0.}

\item{NUMREPS}{\code{integer} number of replicate runs. Defaults to 1L.}

\item{NTHREADS}{\code{integer} number of cores to use for processing. Defaults to 1L.}

\item{GAP}{\code{numeric} MIP gap (lp_solve parameter). Defaults to 0.05.}

\item{PRESOLVE}{\code{integer} code for level of computation in presolve (lp_solve parameter). Defaults to 2L.}

\item{TIMELIMIT}{\code{integer} number of seconds to allow for solving. Defaults to NA_integer_, and so a time limit is not imposed.}

\item{VERBOSITY}{\code{integer} code for amount of output to display. Defaults to 1L.}
}
\value{
\code{MarxanOpts} object
}
\description{
This function creates a new RaspOpts object.
}
\examples{
x<-RaspOpts(NTHREADS=2, NUMREPS=2)
}
\seealso{
\code{\link{RaspOpts-class}}.
}

