% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspOpts.R
\name{RaspOpts}
\alias{RaspOpts}
\title{Create RaspOpts object}
\usage{
RaspOpts(BLM = 0, FAILUREMULTIPLIER = 1000, MAXRLEVEL = 5L,
  NUMREPS = 1L)
}
\arguments{
\item{BLM}{\code{numeric} boundary length modifier. Defaults to 0.}

\item{MAXRLEVEL}{\code{numeric} maximum R failure level for approximation. Defaults to 5L.}

\item{NUMREPS}{\code{integer} number of replicate runs. Defaults to 1L.}
}
\value{
\code{MarxanOpts} object
}
\description{
This function creates a new RaspOpts object.
}
\section{Slots}{

\describe{
\item{\code{FAILUREMULTIPLIER}}{\code{numeric} multiplier for failure planning unit. Defaults to 1000.}
}}
\examples{
x<-RaspOpts(NTHREADS=2, NUMREPS=2)
}
\seealso{
\code{\link{RaspOpts-class}}.
}

