% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R
\docType{class}
\name{RaspResults-class}
\alias{RaspResults-class}
\title{RaspResults: An S4 class to represent RASP results}
\description{
This class is used to store RASP results.
}
\section{Slots}{

\describe{
\item{\code{summary}}{\code{data.frame} with summary information on solutions.}

\item{\code{selections}}{\code{matrix} with binary selections.}

\item{\code{amount.held}}{\code{matrix} with the amount held for each species in each solution.}

\item{\code{space.held}}{\code{matrix} with the poportion of attribute space sampled for each species in each solution.}

\item{\code{best}}{\code{integer} with index of best solution.}

\item{\code{model.file}}{\code{character} with Gurobi model file.}

\item{\code{logging.file}}{\code{character} with Gurobi log file.}

\item{\code{solution.file}}{\code{character} with Gurobi solution file.}

\item{\code{.cache}}{\code{environment} used to store extra data.}
}}
\seealso{
\code{\link{RaspResults}}, \code{\link{read.RaspResults}}.
}

