% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspSolved.R
\docType{class}
\name{RaspSolved-class}
\alias{RaspSolved-class}
\title{RaspSolved: An S4 class to represent RASP inputs and outputs}
\description{
This class is used to store RASP input and output data in addition to input parameters.
}
\section{Slots}{

\describe{
\item{\code{opts}}{\code{RaspReliableOpts} or \code{RaspUnreliableOpts} object used to store input parameters.}

\item{\code{sovler}}{\code{GurobiOpts} object used to store Gurobi parameters.}

\item{\code{data}}{\code{RaspData} object used to store input data.}

\item{\code{results}}{\code{RaspResults} object used to store results.}
}}
\seealso{
\code{\link{RaspReliableOpts-class}}, \code{\link{RaspUnreliableOpts-class}}, \code{\link{RaspData-class}}, \code{\link{RaspResults-class}}.
}

