% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspSolved.R
\name{RaspSolved}
\alias{RaspSolved}
\title{Create new RaspSolved object}
\usage{
RaspSolved(unsolved, solver, results)
}
\arguments{
\item{unsolved}{\code{RaspUnsolved} object.}

\item{solver}{\code{GurobiOpts} or \code{ManualOpts} object.}

\item{results}{\code{RaspResults} object.}
}
\value{
\code{RaspSolved} object.
}
\description{
This function creates a \code{RaspSolved} object.
}
\seealso{
\code{\link{RaspSolved-class}}, \code{\link{RaspResults-class}}, \code{link{solve}}.
}

