% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspSolved.R
\name{RaspSolved}
\alias{RaspSolved}
\title{Create new RaspSolved object}
\usage{
RaspSolved(unsolved, results)
}
\arguments{
\item{unsolved}{\code{RaspUnsolved} object.}

\item{results}{\code{RaspResults} object.}
}
\value{
\code{RaspSolved} object.
}
\description{
This function creates a \code{RaspSolved} object using a \code{RaspUnsolved} object and a \code{RaspResults} object.
}
\examples{
\dontrun{
# load data
data(sim_ru)
# make RaspSolved object
sim_rs <- raspr::solve(sim_ru)
}
}
\seealso{
\code{\link{RaspSolved-class}}, \code{\link{RaspResults-class}}.
}

