% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspUnsolved.R
\name{RaspUnsolved}
\alias{RaspUnsolved}
\title{Create a new RaspUnsolved object}
\usage{
RaspUnsolved(opts, data)
}
\arguments{
\item{opts}{\code{RaspReliableOpts} or \code{RaspUnreliableOpts} object.}

\item{data}{\code{RaspData} object.}
}
\value{
\code{RaspUnsolved} object.
}
\description{
This function creates a \code{RaspUnsolved} object using a \code{GurobiOpts}, a \code{RaspReliableOpts} or \code{RaspUnreliableOpts} object, and a \code{RaspData} object.
}
\examples{
# load data
data(cs_pus, cs_spp)
# create inputs for RaspUnsolved
ro <- RaspUnreliableOpts()
rd <- make.RaspData(cs_pus[1:10,], cs_spp, NULL, include.geographic.space=TRUE,n.demand.points=5L)
# create RaspUnsolved object
ru <- RaspUnsolved(ro, rd)
print(ru)
}
\seealso{
\code{\link{RaspReliableOpts-class}}, \code{\link{RaspUnreliableOpts-class}}, \code{\link{RaspData-class}}.
}

