% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R
\name{read.RaspResults}
\alias{read.RaspResults}
\title{Read RASP results}
\usage{
read.RaspResults(data, path)
}
\arguments{
\item{data}{\code{RaspData} object containing RASP data.}

\item{path}{\code{character} with file path to the output from Gurobi.}
}
\value{
\code{RaspResults} object
}
\description{
This function reads files output from Gurobi and returns a \code{RaspResults} object.
}
\seealso{
\code{\link{RaspResults-class}}, \code{\link{RaspResults}}.
}

