% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R
\name{read.RaspResults}
\alias{read.RaspResults}
\title{Read RASP results}
\usage{
read.RaspResults(opts, data, model.file, log.file, solution.file)
}
\arguments{
\item{opts}{\code{RaspOpts} object}

\item{data}{\code{RaspData} object}

\item{model.file}{\code{character} object containing Gurobi model file.}

\item{log.file}{\code{character} object containing Gurobi log file.}

\item{solution.file}{\code{character} object containing Gurobi solution file.}
}
\value{
\code{RaspResults} object
}
\description{
This function reads files output from Gurobi and returns a \code{RaspResults} object.
}
\seealso{
\code{\link{RaspData}}, \code{\link{RaspData-class}}, \code{\link{RaspResults-class}}, \code{\link{RaspResults}}.
}

