% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/raspr-internal.R
\name{read.RaspResults}
\alias{read.RaspResults}
\title{Read RASP results}
\usage{
read.RaspResults(opts, data, model.list, logging.file, solution.list,
  verbose = FALSE)
}
\arguments{
\item{opts}{\code{RaspReliableOpts} or \code{RaspUnreliableOpts} object}

\item{data}{\code{RaspData} object}

\item{model.list}{\code{list} object containing Gurobi model data}

\item{logging.file}{\code{character} Gurobi log files.}

\item{solution.list}{\code{list} object containing Gurobi solution data.}

\item{verbose}{\code{logical} print progress messages? Defaults to \code{FALSE}.}
}
\value{
\code{RaspResults} object
}
\description{
This function reads files output from Gurobi and returns a \code{RaspResults} object.
}
\seealso{
\code{\link{RaspReliableOpts}}, \code{\link{RaspUnreliableOpts}}, \code{\link{RaspData}}, \code{\link{RaspResults}}.
}
\keyword{internal}

