% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GurobiOpts.R, R/ManualOpts.R, R/RaspData.R, R/RaspReliableOpts.R, R/RaspResults.R, R/RaspSolved.R, R/RaspUnreliableOpts.R, R/RaspUnsolved.R, R/generics.R
\docType{methods}
\name{show,GurobiOpts-method}
\alias{show}
\alias{show,GurobiOpts-method}
\alias{show,ManualOpts-method}
\alias{show,RaspData-method}
\alias{show,RaspReliableOpts-method}
\alias{show,RaspResults-method}
\alias{show,RaspSolved-method}
\alias{show,RaspUnreliableOpts-method}
\alias{show,RaspUnsolved-method}
\title{Show objects}
\usage{
\S4method{show}{GurobiOpts}(object)

\S4method{show}{ManualOpts}(object)

\S4method{show}{RaspData}(object)

\S4method{show}{RaspReliableOpts}(object)

\S4method{show}{RaspResults}(object)

\S4method{show}{RaspSolved}(object)

\S4method{show}{RaspUnreliableOpts}(object)

\S4method{show}{RaspUnsolved}(object)
}
\arguments{
\item{object}{\code{GurobiOpts}, \code{RaspUnreliableOpts}, \code{RaspReliableOpts}, \code{RaspData}, \code{RaspUnsolved}, \code{RaspResults}, or \code{RaspSolved} object.}
}
\description{
Shows objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{GurobiOpts}: 

\item \code{ManualOpts}: 

\item \code{RaspData}: 

\item \code{RaspReliableOpts}: 

\item \code{RaspResults}: 

\item \code{RaspSolved}: 

\item \code{RaspUnreliableOpts}: 

\item \code{RaspUnsolved}: 
}}
\examples{
# load data
data(sim_ru, sim_rs)
## show methods for classes in package
# GurobiOpts
GurobiOpts()
# RaspReliableOpts
RaspReliableOpts()
# RaspUnreliableOpts
RaspUnreliableOpts()
# RaspData
sim_ru@data
# RaspUnsolved
sim_ru
# RaspResults
sim_rs@results
# RaspSolved
sim_rs
}
\seealso{
\code{GurobiOpts}, \code{RaspUnreliableOpts}, \code{RaspReliableOpts}, \code{RaspData}, \code{RaspUnsolved}, \code{RaspResults}, \code{RaspSolved}.
}

