% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/GurobiOpts.R, R/ManualOpts.R,
%   R/RapData.R, R/RapReliableOpts.R, R/RapResults.R, R/RapUnreliableOpts.R,
%   R/RapUnsolved.R, R/RapSolved.R
\docType{methods}
\name{show}
\alias{show}
\alias{show}
\alias{show,GurobiOpts-method}
\alias{show}
\alias{show,ManualOpts-method}
\alias{show}
\alias{show}
\alias{show}
\alias{show}
\alias{show}
\alias{show}
\title{Show objects}
\usage{
\S4method{show}{GurobiOpts}(object)

\S4method{show}{ManualOpts}(object)
}
\arguments{
\item{object}{\code{\link{GurobiOpts}}, \code{\link{RapUnreliableOpts}},
\code{\link{RapReliableOpts}}, \code{\link{RapData}},
\code{\link{RapUnsolved}}, \code{\link{RapResults}}, or
\code{\link{RapSolved}} object.}
}
\description{
Shows objects.
}
\examples{
# load data
data(sim_ru, sim_rs)

# show GurobiOpts object
GurobiOpts()

# show RapReliableOpts object
RapReliableOpts()

# show RapUnreliableOpts object
RapUnreliableOpts()

# show RapData object
sim_ru@data

# show RapUnsolved object
sim_ru

# show RapResults object
sim_rs@results

# show RapSolved object
sim_rs
}
\seealso{
\code{\link{GurobiOpts}}, \code{\link{RapUnreliableOpts}},
  \code{\link{RapReliableOpts}}, \code{\link{RapData}},
  \code{\link{RapUnsolved}}, \code{\link{RapResults}},
  \code{\link{RapSolved}}.
}
