% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GurobiOpts.R, R/ManualOpts.R, R/RapData.R, R/RapReliableOpts.R, R/RapResults.R, R/RapSolved.R, R/RapUnreliableOpts.R, R/RapUnsolved.R, R/generics.R
\docType{methods}
\name{show,GurobiOpts-method}
\alias{show}
\alias{show,GurobiOpts-method}
\alias{show,ManualOpts-method}
\alias{show,RapData-method}
\alias{show,RapReliableOpts-method}
\alias{show,RapResults-method}
\alias{show,RapSolved-method}
\alias{show,RapUnreliableOpts-method}
\alias{show,RapUnsolved-method}
\title{Show objects}
\usage{
\S4method{show}{GurobiOpts}(object)

\S4method{show}{ManualOpts}(object)

\S4method{show}{RapData}(object)

\S4method{show}{RapReliableOpts}(object)

\S4method{show}{RapResults}(object)

\S4method{show}{RapSolved}(object)

\S4method{show}{RapUnreliableOpts}(object)

\S4method{show}{RapUnsolved}(object)
}
\arguments{
\item{object}{\code{GurobiOpts}, \code{RapUnreliableOpts}, \code{RapReliableOpts}, \code{RapData}, \code{RapUnsolved}, \code{RapResults}, or \code{RapSolved} object.}
}
\description{
Shows objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{GurobiOpts}: 

\item \code{ManualOpts}: 

\item \code{RapData}: 

\item \code{RapReliableOpts}: 

\item \code{RapResults}: 

\item \code{RapSolved}: 

\item \code{RapUnreliableOpts}: 

\item \code{RapUnsolved}: 
}}
\examples{
# load data
data(sim_ru, sim_rs)
## show methods for classes in package
# GurobiOpts
GurobiOpts()
# RapReliableOpts
RapReliableOpts()
# RapUnreliableOpts
RapUnreliableOpts()
# RapData
sim_ru@data
# RapUnsolved
sim_ru
# RapResults
sim_rs@results
# RapSolved
sim_rs
}
\seealso{
\code{\link{GurobiOpts}}, \code{\link{RapUnreliableOpts}}, \code{\link{RapReliableOpts}}, \code{\link{RapData}}, \code{\link{RapUnsolved}}, \code{\link{RapResults}}, \code{\link{RapSolved}}.
}

