% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspSolved.R, R/RaspUnsolved.R, R/generics.R
\docType{methods}
\name{solve,RaspSolved-method}
\alias{solve}
\alias{solve,RaspSolved-method}
\alias{solve,RaspUnsolved-method}
\title{Solve RASP object}
\usage{
\S4method{solve}{RaspSolved}(x, verbose = FALSE, force.reset = FALSE)

\S4method{solve}{RaspUnsolved}(x, verbose = FALSE)

solve(x, ...)
}
\arguments{
\item{x}{\code{RaspUnsolved} or \code{RaspSolved} object.}

\item{verbose}{\code{logical} should messages be printed during creation of the initial model matrix?.}

\item{force.reset}{\code{logical} if \code{x} already has solutions, should this be discarded and a new solution(s) obtained?}

\item{...}{not used.}
}
\value{
\code{RaspSolved} object
}
\description{
This function uses Gurobi to find prioritisations using the input parameter and data stored in a \code{RaspUsolved} object,
and returns a \code{RaspSolved} object with outputs in it.
}
\section{Methods (by class)}{
\itemize{
\item \code{RaspSolved}: 

\item \code{RaspUnsolved}: 
}}
\note{
This function is used to solve a \code{RaspUnsolved} object that has all of its inputs generated. The rasp function (without lower case 'r') provides a more general interface for generating inputs and outputs.
}
\seealso{
\code{RaspUnsolved}, \code{RaspSolved}.
}

