% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspResults.R, R/RaspSolved.R, R/generics.R
\name{space.held.RaspResults}
\alias{space.held}
\alias{space.held.RaspResults}
\alias{space.held.RaspSolved}
\title{Extract attribute space held for a solution}
\usage{
\method{space.held}{RaspResults}(x, y = 0)

\method{space.held}{RaspSolved}(x, y = 0)

space.held(x, y)
}
\arguments{
\item{x}{\code{RaspResults} or \code{RaspSolved} object.}

\item{y}{\code{NULL} to return all values, \code{integer} 0 to return values for best solution, \code{integer} value greater than 0 for \code{y}'th solution value.}
}
\value{
code{matrix} or code{numeric} vector depending on arguments.
}
\description{
This function returns the attribute space held for each species in a solution.
}
\examples{
load(sim_rs)
# space held (\%) for each species in best solution
space.held(sim_rs, 0)
# space held (\%) for each species in second solution
space.held(sim_rs, 2)
# space held (\%) for each species in each solution
space.held(sim_rs, NULL)
}
\seealso{
\code{RaspResults}, \code{RaspSolved}.
}

