% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RapData.R, R/RapSolved.R, R/RapUnsolved.R, R/generics.R
\name{space.plot.RapData}
\alias{space.plot}
\alias{space.plot.RapData}
\alias{space.plot.RapSolved}
\alias{space.plot.RapUnsolved}
\title{Plot space}
\usage{
\method{space.plot}{RapData}(x, species, space = 1,
  pu.color.palette = "RdYlGn", locked.in.color = "#000000FF",
  locked.out.color = "#D7D7D7FF", ...)

\method{space.plot}{RapSolved}(x, species, space = 1, y = 0,
  pu.color.palette = "RdYlGn", locked.in.color = "#000000FF",
  locked.out.color = "#D7D7D7FF", ...)

\method{space.plot}{RapUnsolved}(x, species, space = 1,
  pu.color.palette = "RdYlGn", locked.in.color = "#000000FF",
  locked.out.color = "#D7D7D7FF", ...)

space.plot(x, species, space, ...)
}
\arguments{
\item{x}{\code{RapData}, \code{RapUnsolved}, or \code{RapSolved} object.}

\item{species}{\code{character} name of species, or \code{integer} index for species.}

\item{space}{\code{integer} index of attribute space.}

\item{pu.color.palette}{\code{character} name of color palette to use for planning units (see \code{\link[RColorBrewer]{brewer.pal}}). Defaults to 'RdYlGn'.}

\item{locked.in.color}{\code{character} color to denote locked in planning units. Used when \code{y} is \code{NULL}. Defaults to '#000000FF'.}

\item{locked.out.color}{\code{character} color to denote locked in planning units. Used when \code{y} is \code{NULL}. Defaults to '#D7D7D7FF'.}

\item{...}{not used.}

\item{y}{\code{NULL} \code{integer} 0 to return values for best solution, \code{integer} value greater than 0 for \code{y}'th solution value.}
}
\description{
This function plots the distribution of planning units and the distribution of demand points for a particular species in an attribute space.
Note that this function only works for attribute spaces with one, two, or three dimensions.
}
\examples{
# load RapSolved objects
data(sim_rs)
# plot distribution of first species in first attribute space
space.plot(sim_rs, 1, 1)
}

