% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R, R/RaspSolved.R, R/generics.R
\name{space.target.RaspData}
\alias{space.target}
\alias{space.target.RaspData}
\alias{space.target.RaspUnsolOrSol}
\alias{space.target<-}
\alias{space.target<-.RaspData}
\alias{space.target<-.RaspUnsolOrSol}
\title{Attribute space targets}
\usage{
\method{space.target}{RaspData}(x, species = NULL, space = NULL)

\method{space.target}{RaspData}(x, species = NULL, space = NULL) <- value

\method{space.target}{RaspUnsolOrSol}(x, species = NULL, space = NULL)

\method{space.target}{RaspUnsolOrSol}(x, species = NULL, space = NULL) <-
  value

space.target(x, species, space)

space.target(x, species, space) <- value
}
\arguments{
\item{x}{\code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.}

\item{species}{\code{NULL} for all species or \code{integer} indicating species.}

\item{space}{\code{NULL} for all spaces or \code{integer} indicating a specific space.}

\item{value}{\code{numeric} new target..}
}
\value{
\code{numeric matrix}.
}
\description{
This function sets or returns the attribute space targets for each species.
}
\examples{
# load data
load(sim_rs)
# extract space targets for all species
space.target(sim_rs)
# set space targets for all species
space.target(sim_rs) <- 0.1
# extract target for first species for first space
space.target(sim_rs, 1, 1)
# set space targets for first species for first space
space.target(sim_rs, 1, 1) <- 0.5
}
\seealso{
\code{RaspData}, \code{RaspResults}, \code{RaspSolved}.
}

