% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RaspData.R, R/RaspSolved.R, R/RaspUnsolved.R, R/generics.R
\name{spp.plot.RaspData}
\alias{spp.plot}
\alias{spp.plot.RaspData}
\alias{spp.plot.RaspSolved}
\alias{spp.plot.RaspUnsolved}
\title{Plot species}
\usage{
\method{spp.plot}{RaspData}(x, y, basemap, color.palette, alpha, grayscale,
  force.reset)

\method{spp.plot}{RaspSolved}(x, y, basemap = "none",
  color.palette = "YlGnBu", alpha = ifelse(basemap == "none", 1, 0.7),
  grayscale = FALSE, force.reset = FALSE)

\method{spp.plot}{RaspUnsolved}(x, y, basemap = "none",
  color.palette = "YlGnBu", alpha = ifelse(basemap == "none", 1, 0.7),
  grayscale = FALSE, force.reset = FALSE)

spp.plot(x, y, basemap, color.palette, alpha, grayscale, force.reset)
}
\arguments{
\item{x}{\code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.}

\item{y}{\code{character} name of species, or \code{integer} index for species.}

\item{basemap}{\code{character} object indicating the type of basemap to use (see \code{link{basemap}}). Use either 'none', 'roadmap', 'mobile', 'satellite', 'terrain', 'hybrid', 'mapmaker-roadmap', 'mapmaker-hybrid'. Defaults to 'none'.}

\item{color.palette}{\code{character} name of colour palette to use for planning units (see \code{\link[RColorBrewer]{brewer.pal}}). Defaults to 'YlGnBu'.}

\item{alpha}{\code{numeric} value to indicate how transparent the planning unit colors shoud be.}

\item{grayscale}{\code{logical} should the basemap be gray-scaled?}

\item{force.reset}{\code{logical} if basemap data has been cached, should it be re-downloaded?}
}
\description{
This function plots the distribution of species across the study area.
}

