% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RapData.R, R/RapSolved.R, R/RapUnsolved.R, R/generics.R
\name{spp.subset.RapData}
\alias{spp.subset}
\alias{spp.subset.RapData}
\alias{spp.subset.RapSolved}
\alias{spp.subset.RapUnsolved}
\title{Subset species}
\usage{
\method{spp.subset}{RapData}(x, species)

\method{spp.subset}{RapSolved}(x, species)

\method{spp.subset}{RapUnsolved}(x, species)

spp.subset(x, species)
}
\arguments{
\item{x}{\code{RapData}, \code{RapUnsolved}, or \code{RapSolved} object.}

\item{species}{\code{integer}, or \code{character} vectors to specify the index or species names to subset.}
}
\value{
\code{RapData} or \code{RapUnsolved} object depending on input object.
}
\description{
Subset species from a \code{RapData}, \code{RapUnsolved}, or \code{RapSolved} object.
}
\examples{
# load data
data(sim_ru)
# generate new object with only species 1
sim_ru2 <- spp.subset(sim_ru, 1)
}
\seealso{
\code{RapData}, \code{RapUnsolved}, \code{RapSolved}.
}

