% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GurobiOpts.R, R/ManualOpts.R, R/RaspData.R, R/RaspReliableOpts.R, R/RaspSolved.R, R/RaspUnreliableOpts.R, R/generics.R
\name{update.GurobiOpts}
\alias{update}
\alias{update.GurobiOpts}
\alias{update.ManualOpts}
\alias{update.RaspData}
\alias{update.RaspReliableOpts}
\alias{update.RaspUnreliableOpts}
\alias{update.RaspUnsolOrSol}
\title{Update object}
\usage{
\method{update}{GurobiOpts}(object, Threads = NULL, MIPGap = NULL,
  Presolve = NULL, TimeLimit = NULL, NumberSolutions = NULL, ...)

\method{update}{ManualOpts}(object, NumberSolutions = NULL, ...)

\method{update}{RaspData}(object, species = NULL, space = NULL,
  name = NULL, amount.target = NULL, space.target = NULL, pu = NULL,
  cost = NULL, status = NULL, ...)

\method{update}{RaspReliableOpts}(object, BLM = NULL,
  FAILUREMULTIPLIER = NULL, MAXRLEVEL = NULL, ...)

\method{update}{RaspUnsolOrSol}(object, ..., solve = TRUE)

\method{update}{RaspUnreliableOpts}(object, BLM = NULL, ...)
}
\arguments{
\item{object}{\code{GurobiOpts}, \code{RaspUnreliableOpts}, \code{RaspReliableOpts}, \code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.}

\item{Threads}{\code{integer} number of cores to use for processing.}

\item{MIPGap}{\code{numeric} MIP gap specifying minimum solution quality.}

\item{Presolve}{\code{integer} code for level of computation in presolve.}

\item{TimeLimit}{\code{integer} number of seconds to allow for solving.}

\item{NumberSolutions}{\code{integer} number of solutions to generate.}

\item{...}{parameters passed to \code{update.RaspReliableOpts}, \code{update.RaspUnreliableOpts}, or \code{update.RaspData}.}

\item{species}{\code{integer} or \code{character} denoting species for which targets or name should be updated.}

\item{space}{\code{integer} denoting space for which targets should be updated.}

\item{name}{\code{character} to rename species.}

\item{amount.target}{\code{numeric} vector for new area targets (\%) for the specified species.}

\item{space.target}{\code{numeric} vector for new attribute space targets (\%) for the specified species and attribute spaces.}

\item{pu}{\code{integer} planning unit indices that need to be updated.}

\item{cost}{\code{numeric} new costs for spcified planning units.}

\item{status}{\code{integer} new statuses for specified planning units.}

\item{BLM}{\code{numeric} boundary length modifier.}

\item{FAILUREMULTIPLIER}{\code{numeric} multiplier for failure planning unit.}

\item{MAXRLEVEL}{\code{numeric} maximum R failure level for approximation.}

\item{solve}{\code{logical} should the problem be solved? This argument is only valid for \code{RaspUnsolved} and \code{RaspSolved} objects. Defaults to \code{TRUE}.}
}
\value{
\code{\link{GurobiOpts-class}}, \code{\link{RaspUnreliableOpts-class}}, \code{\link{RaspReliableOpts-class}}, \code{\link{RaspData-class}}, \code{\link{RaspUnsolved-class}}, or \code{\link{RaspSolved-class}} object depending on \code{x}.
}
\description{
This function updates parameters or data stored in an existing \code{GurobiOpts}, \code{RaspUnreliableOpts}, \code{RaspReliableOpts}, \code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object.
}
\examples{
# load data
data(sim_ru, sim_rs)

# GurobiOpts
x <- GurobiOpts(MIPGap=0.7)
y <- update(x, MIPGap=0.1)
print(x)
print(y)

# RaspUnreliableOpts
x <- RaspUnreliableOpts(BLM=10)
y <- update(x, BLM=2)
print(x)
print(y)

# RaspReliableOpts
x <- RaspReliableOpts(FAILUREMULTIPLIER=2)
y <- update(x, FAILUREMULTIPLIER=4)
print(x)
print(y)

# RaspData
x <- sim_ru@data
y <- update(x, space.target=c(0.4, 0.7, 0.1))
print(space.target(x))
print(space.target(y))

## RaspUnsolved
x <- sim_ru
y <- update(x, amount.target=c(0.1, 0.2, 0.3), BLM=3, solve=FALSE)
# print x parameters
print(x@opts@BLM); print(amount.target(x))
# print y parameters
print(y@opts@BLM); print(space.target(y))

## RaspSolved
x <- sim_rs
y <- update(x, space.targets=c(0.4, 0.6, 0.9), BLM=100, Presolve=1L, solve=FALSE)
# print x parameters
print(x@opts@BLM); print(amount.target(x))
# print y parameters
print(y@opts@BLM); print(space.target(y))
}
\seealso{
\code{\link{GurobiOpts-class}}, \code{\link{RaspUnreliableOpts-class}}, \code{\link{RaspReliableOpts-class}}, \code{\link{RaspData-class}}, \code{\link{RaspUnsolved-class}}, \code{\link{RaspSolved-class}}.
}

