% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GurobiOpts.R, R/RaspData.R, R/RaspOpts.R, R/RaspSolved.R, R/RaspUnsolved.R, R/generics.R
\name{update.GurobiOpts}
\alias{update}
\alias{update.GurobiOpts}
\alias{update.RaspData}
\alias{update.RaspOpts}
\alias{update.RaspSolved}
\alias{update.RaspUnsolved}
\title{Update object}
\usage{
\method{update}{GurobiOpts}(object, ...)

\method{update}{RaspData}(object, ...)

\method{update}{RaspOpts}(object, ...)

\method{update}{RaspSolved}(object, ...)

\method{update}{RaspUnsolved}(object, ...)
}
\arguments{
\item{object}{\code{GurobiOpts}, \code{RaspOpts}, \code{RaspData}, \code{RaspUnsolved}, or \code{RaspSolved} object..}

\item{...}{parameters to update.}
}
\value{
\code{GurobiOpts}, \code{RaspOpts}, \code{RaspData}, or \code{RaspUnsolved} object depending on \code{x}.
}
\description{
This function updates parameters or data stored in an existing
\code{GurobiOpts}, \code{RaspOpts}, \code{RaspData},
\code{RaspUnsolved}, or \code{RaspSolved} object.
}
\seealso{
\code{GurobiOpts}, \code{RaspOpts}, \code{RaspData}, or \code{RaspUnsolved}.
}

